\name{addCaption}
\alias{addCaption}
\title{
Add Caption
}
\description{
Adds a caption at the bottom of the graph
}
\usage{
addCaption(caption='')
}
\arguments{
  \item{caption}{
the text of the caption for the graph
}
}
\value{
Nothing is returned.
}
\note{
Useful for adding captions when the explanation is at the bottom.
}
\keyword{ aplot }

