% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExplanation.R
\name{addExplanation}
\alias{addExplanation}
\title{Add Explanation}
\usage{
addExplanation(what, where = "new",
  title = expression(bold(EXPLANATION)), box.off = where != "new",
  margin = rep(0, 4), line.length = 2, line.height = 1.1)
}
\arguments{
\item{what}{a specialized object for an explanation, from the output from
calls to the plotting functions}

\item{where}{a description of where to put the explanation, see
\bold{Details}.}

\item{title}{the title of the explanation.}

\item{box.off}{logical, if \code{TRUE}, then box off the explanation with a
blank background and black box, otherwise the background is not blanked and no
bounding box is drawn.}

\item{margin}{the margin for a new graph}

\item{line.length}{the relative length of lines drawn in the explanation, see
\bold{Details}.}

\item{line.height}{the relative spacing of the lines in the explanation.}
}
\value{
Nothing is returned.
}
\description{
Creates or adds an explanation, also called key or legend.
}
\details{
The value for \code{where} must be one of "ul," "ur," "ll," "lr," 
"cl," "cr," "uc," "lc,"
"cc," or "new." If "new," then the explanation is placed in a new graph,
otherwise, the first letter is an abbreviation for upper, lower, or center
and the second letter is an abbreviation for left, right, or center. the
explanation for a box plot, stiff diagram, or contour plot must be placed 
in a new graph. If \code{box.off} is \code{TRUE}, then the explanation abuts
the axes, otherwise it is placed slightly inset so that the text does not
interfere with the ticks.

In most cases, \code{line.length} does not need to be changed. In some cases,
such as mass produced figures that will not be modified by an illustrator, the
\code{line.length} can be increased to show full dashes if dashed lines are
drawn. In gneralt, the illustrator should create dashed lines rather than drawing
them in the graph.
}
\note{
The call to \code{addExplanation} should be the last in any sequence of
calls to construct a figure becuase it can alter some graphical parameters.\cr
Box plot explanations require fairly large graph areas because of the
detail required for some types. In general, a graph about 4.5 inches high is
needed for the Tukey type and 4 inches for other types and widths of 2.5 and
2 inches respectively. The sizes are smaller for the font type of "USGS."
If the graph area is smaller than required for the box plot explanation, then 
either a modified explanation is created or a warning is printed and the 
explanation may be unreadable.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
setGD()
AA.pl <- xyPlot(X, Y, Plot=list(name="Random Points"))
addExplanation(AA.pl, where='ul')
# For more details of addExplanation see
vignette(topic="BoxPlots", package="smwrGraphs")
vignette(topic="GraphAdditions", package="smwrGraphs")
vignette(topic="GraphGallery", package="smwrGraphs")
vignette(topic="GraphSetup", package="smwrGraphs")
vignette(topic="LineScatter", package="smwrGraphs")
vignette(topic="PiperPlot", package="smwrGraphs")
demo(topic="Coplot-complexScatterPlot", package="smwrGraphs")
demo(topic="FlowDur-Measurements", package="smwrGraphs")
demo(topic="PiperScript", package="smwrGraphs")
demo(topic="RightAxisExample", package="smwrGraphs")
}
}
\keyword{aplot}
