% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addSLR}
\alias{addSLR}
\alias{addSLR.default}
\alias{addSLR.list}
\title{Add a Regression Line}
\usage{
addSLR(x, y, Plot = list(name = "", what = "lines", type = "solid", width =
  "standard", color = "black"), Model = list(x = "", y = "", form = "exp",
  where = "none"), current = list(yaxis.log = FALSE, yaxis.rev = FALSE,
  xaxis.log = FALSE), ...)

\method{addSLR}{default}(x, y, Plot = list(name = "", what = "lines", type =
  "solid", width = "standard", color = "black"), Model = list(x = "", y = "",
  form = "exp", where = "none"), current = list(yaxis.log = FALSE, yaxis.rev =
  FALSE, xaxis.log = FALSE), ...)

\method{addSLR}{list}(x, y, Plot = list(name = "", what = "lines", type =
  "solid", width = "standard", color = "black"), Model = list(x = "", y = "",
  form = "exp", where = "none"), current = list(yaxis.log = FALSE, yaxis.rev =
  FALSE, xaxis.log = FALSE), ...)
}
\arguments{
\item{x}{the x-axis data. For method \code{list}, x is a list that contains
components \code{x} and \code{y} and the \code{y} argument is not used.
Missing values are permitted but ignored.}

\item{y}{the y-axis data. Missing values are permitted but ignored.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{current}{the current plot information. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}

\item{\dots}{not used required for other methods.}
}
\value{
The current plot information, the x and y components are the data, not the line.
The regression model is included as the lm component.
}
\description{
Computes a simple linear regression from x and y data and adds the line to a graph.
}
\seealso{
\code{\link{addXY}}, \code{\link{xyPlot}}
}
\keyword{aplot}

