% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addSmooth.R
\name{addSmooth}
\alias{addSmooth}
\alias{addSmooth.default}
\alias{addSmooth.list}
\title{Add a Smooth Line}
\usage{
addSmooth(x, y, Smooth = "loess.smooth", ..., Smooth.along = "x",
  Plot = list(name = "", what = "lines", type = "solid", width = "standard",
  color = "black"), current = list(yaxis.log = FALSE, yaxis.rev = FALSE,
  xaxis.log = FALSE))

\method{addSmooth}{default}(x, y, Smooth = "loess.smooth", ...,
  Smooth.along = "x", Plot = list(name = "", what = "lines", type = "solid",
  width = "standard", color = "black"), current = list(yaxis.log = FALSE,
  yaxis.rev = FALSE, xaxis.log = FALSE))

\method{addSmooth}{list}(x, y, Smooth = "loess.smooth", ...,
  Smooth.along = "x", Plot = list(name = "", what = "lines", type = "solid",
  width = "standard", color = "black"), current = list(yaxis.log = FALSE,
  yaxis.rev = FALSE, xaxis.log = FALSE))
}
\arguments{
\item{x}{the x-axis data. For method \code{list}, x is a list that contains
components \code{x} and \code{y} and the \code{y} argument is not used.
Missing values are permitted and ignored in the smooth.}

\item{y}{the y-axis data. Missing values are permitted and ignored in the smooth.}

\item{Smooth}{the name of the smoothing function.}

\item{Smooth.along}{the data along which the smoother is run. Must be either "x,"
which smooths \code{y} along \code{x} resulting in a horzontal line, or "y,"
which smooths \code{x} along \code{y} resulting in a vereical line.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{current}{the current plot information. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}

\item{\dots}{additional parameters for the function names in \code{Smooth}.}
}
\value{
The current plot information.
}
\description{
Computes a smooth line from x and y data and adds the line to a graph.
}
\note{
If an error is generated from the smoother, then nothing is added to the
graph, an error is printed, the returned object contains missing values for
the data that should have been plotted, and the explanation is not updated.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
AA.pl <- xyPlot(X, Y)
addSmooth(AA.pl)
# For more details of addSmooth see
vignette(topic="GraphAdditions", package="smwrGraphs")
}
}
\seealso{
\code{\link{addXY}}, \code{\link{xyPlot}}
}
\keyword{aplot}

