\name{addTitle}
\alias{addTitle}
\title{
Add a Title
}
\description{
Adds a title to a graph.
}
\usage{
addTitle(Main = "", Justification = "left", Bold = FALSE)
}
\arguments{
  \item{Main}{
the title.
}
  \item{Justification}{
specifies the location of the title, must be one of "left," "center," or "right."
}
  \item{Bold}{
logical: should the title be in bold face type?
}
}
\value{
Nothing is returned.
}
\seealso{
\code{\link{addCaption}}, \code{\link{addAnnotation}},  \code{\link{addTable}}
}
\keyword{aplot}

