% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{addXY}
\alias{addXY}
\alias{addXY,ANY,numeric-method}
\alias{addXY,numeric,character-method}
\title{Add a plot to a graph}
\usage{
addXY(x, y, ...)

\S4method{addXY}{ANY,numeric}(x, y, Plot = list(name = "", what = "lines",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "black"), current = list(yaxis.log = FALSE, yaxis.rev = FALSE,
  xaxis.log = FALSE), new.axis = "none", new.log = FALSE, new.rev = FALSE,
  new.range = c(NA, NA), new.labels = 7, new.title = "")

\S4method{addXY}{numeric,character}(x, y, Plot = list(name = "", what =
  "points", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "black"), current = list(yaxis.log = FALSE,
  yaxis.rev = FALSE, xaxis.log = FALSE), jitter.y = FALSE)
}
\arguments{
\item{x}{the x-axis data. Missing values are permitted, but result in breaks
in the plotted data.}

\item{y}{the y-axis data. Missing values are permitted, but result in breaks
in the plotted data.}

\item{...}{arguments for specific methods.}

\item{Plot}{parameters of the plot.  See \code{\link{setPlot}} for a
description of the parameters.}

\item{current}{the current plot information. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}

\item{new.axis}{character:  indicating which new axis to set up. Must be either "right," "top,"
or "none," which indicates that the existing axes be used (default).}

\item{new.log}{logical: log transform new axis?}

\item{new.rev}{logical: reverse new axis?}

\item{new.range}{set new-axis range.}

\item{new.labels}{set up new-axis labels.}

\item{new.title}{the new-axis title.}

\item{jitter.y}{adjust \code{y} values to reduce overlap for each group?}
}
\value{
Information about the graph.
}
\description{
Add points or lines to the current graph.
}
\section{Methods}{
 \describe{

\item{signature(x ="ANY", y = "numeric"}{Any valid x-axis data and numeric y.}
\item{signature(x ="numeric", y = "character"}{Method to add to a dot plot;
the right-axis arguments are not valid.}
}
}
\keyword{aplot}
\keyword{methods}

