% Generated by roxygen2 (4.0.2): do not edit by hand
\name{areaPlot}
\alias{areaPlot}
\title{Shaded Area Plot}
\usage{
areaPlot(x, y, Areas = list(name = "Auto", fillDir = "between", base = "Auto",
  lineColor = "black", fillColors = "pastel"), yaxis.log = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = "Auto", xtitle = "", ytitle = "",
  caption = "", margin = c(NA, NA, NA, NA))
}
\arguments{
\item{x}{x-axis coordinates in increasing order.}

\item{y}{a numeric matrix of y-axis coordinates.}

\item{Areas}{parameters controlling the areas. See \bold{Details}.}

\item{yaxis.log}{logical: log transform y axis?}

\item{yaxis.range}{set y-axis range.}

\item{xaxis.log}{logical: log transform x axis?}

\item{xaxis.range}{set x-axis range.}

\item{ylabels}{set up y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set up x-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the caption for the graph.}

\item{margin}{set up the plot area margins.}
}
\description{
Produces an plot where the area between lines is filled with color.
}

