% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biPlot.R
\name{biPlot}
\alias{biPlot}
\title{Biplot}
\usage{
biPlot(x, ...)
}
\arguments{
\item{x}{any object which has a valid method for \code{biPlot}.}

\item{\dots}{additional arguments for other methods.}
}
\value{
Information about the graph
}
\description{
Produce a biplot, which is a plot of two different types of data on the same
graph.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{biPlot}.
}
\examples{
\dontrun{
# See for examples of biPlot:
vignette(topic="GraphGallery", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{biPlot.default}},
\code{\link{biPlot.princomp}}
}
\keyword{hplot}

