% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Conditioned Graphs}
\usage{
condition(plot.call, data, group, format = "grid", num.cols = NA,
  num.rows = NA, explanation = NULL, share = "", group.name = "",
  xtitle = "", ytitle = "", caption = "")
}
\arguments{
\item{plot.call}{either a simple call to a graphics function or a sequence
of calls enclosed in curly braces(\{\})}

\item{data}{the data.frame containing the variables used in \code{plot.call}
and \code{group}.}

\item{group}{a character string identifying the grouping variable in
\code{data}.}

\item{format}{the orientation of the graphs. If "table," then the graphs are
created beginning in the upper-left hand corner. If "grid," then the graphs
are created beginning in the lower-left See \bold{Details}. hand corner.}

\item{num.cols}{the number of columns on each page. See \bold{Details}.}

\item{num.rows}{the number of rows on each page. See \bold{Details}.}

\item{explanation}{where to place an explanation if necessary. See
\code{\link{setLayout}} for details.}

\item{share}{character, if \code{share} contains "x", then the code in 
\code{plot.call} is set up to share x-axes; if \code{share} contains "y", 
then the code in \code{plot.call} is set up to share y-axes. The default is
not to share either axes.}

\item{group.name}{a character string to prepend to each value in
\code{group} to create the graph title.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}
}
\value{
The value returned by \code{plot.call} is returned invisibly. If
used correctly, it could be used to add an explanation.
}
\description{
Facilitates producing a series of graphs conditioned by a grouping variable.
}
\details{
Graphs are created by executing \code{plot.call} for each unique value in
the variable specified by \code{group}. The columns and rows are set by
\code{num.cols} and \code{num.rows}. If both are set to missing \code{NA},
then each page contains 1 graph, and the graphics setup should provide for
mulitple pages. If one of \code{num.cols} or \code{num.rows} is set to a
numeric value, the other is calculated from the number of graphs so that all
graphs would be displayed on a single page. If both \code{num.cols} and
\code{num.rows} are set to numeric values, then each page would contain a
maximum of \code{num.cols} times \code{num.rows} and multiple pages would be
needed if that were less than the total number of graphs.\cr

The order of the graphs is controlled by the type of \code{group}. If
\code{group} is a factor, then the order is set by the order of the levels,
otherwise the order is set by the order from \code{unique}.
}
\note{
This function is designed to facilitate the production of conditioned
graphs, but may not render each collection completely. For example, graphs
arranged in table format will not have x-axis labels for incomplete
columns.\cr

A call must be made to \code{setPage} or \code{setPDF} to set up the
graphics page before calling \code{condition}. The returned value can be
used to create an explanation, if desired. If \code{explanation} is not
\code{NULL}, then the graph is set to the explanation and there is no need
to call \code{setGraph} to set up the explanation.\cr

The called plotting function must set the \code{margin} argument to
\code{.margin}. See the demos for examples.
}
\examples{
\dontrun{
# See for examples of condition:
demo(topic="Coplot-complexScatterPlot", package="smwrGraphs")
demo(topic="Coplot-simpleBoxPlot", package="smwrGraphs")
}
}
\references{
Cleveland, W.S., 1993, Visualizing data: Summit, New Jersey,
Hobart Press, 360 p.
}
\seealso{
\code{\link{setLayout}}, \code{\link{setPage}},
\code{\link{setPDF}}, \code{\link{unique}}
}
\keyword{hplot}
