% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corGram.R
\name{corGram}
\alias{corGram}
\title{Correlogram}
\usage{
corGram(x, y, Plot = list(name = "Standardized Observations", what = "points",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.03, color = "gray40"), CorGram = list(band = 0.15, kernel = "normal",
  color = "black", width = "standard", add0line = TRUE), yaxis.range = c(-3,
  3), xaxis.range = c(0, 1.5), ylabels = 7, xlabels = 4,
  xtitle = "Difference in Time, in Years",
  ytitle = "Standardized Serial Correlation", caption = "", margin = c(NA,
  NA, NA, NA))
}
\arguments{
\item{x}{decimal time.}

\item{y}{residuals or other observations, these will be scaled, but not
centered.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{CorGram}{control parameters of the correlogram line. See \bold{Details}.}

\item{yaxis.range}{set the y-axis range.}

\item{xaxis.range}{set the x-axis range.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set the x-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}
}
\value{
Information about the graph.
}
\description{
Creates a correlogram for irregularly spaced data.
}
\details{
\code{CorGram} is a list with these components: 
\describe{ 
\item{band}{a measure of the bandwidth used by \code{\link[KernSmooth]{locpoly}}}
\item{kernel}{the kernel used, currently ignored}
\item{color}{the color of the line representing the correlogram}
\item{width}{the line width of the line representing the correlogram}
\item{add0line}{logical, if \code{TRUE}, then add the 0 line; if \code{FALSE}, 
then do not draw the 0 line}}
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{corGram}.
}
\examples{
\dontrun{
# See for examples of corGram:
vignette(topic="GraphGallery", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}
}
\keyword{hplot}
