% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dendGram.R
\name{dendGram}
\alias{dendGram}
\title{Tree Graphs}
\usage{
dendGram(x, Tree = list(orientation = "vertical", width = "standard", color =
  "black"), axis.range = c(NA, NA), labels = "Auto", ytitle = "",
  xtitle = "", caption = "", margin = c(NA, NA, NA, NA))
}
\arguments{
\item{x}{the data to plot. Must be able to be converted to class
"dendrogram."}

\item{Tree}{control parameters of the tree diagram.}

\item{axis.range}{set the range of the tree-axis.}

\item{labels}{set the tree-axis labels. See \code{\link{linearPretty}} for
details.}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{the parameters of the margin of the plot area.}
}
\value{
Information about the graph.
}
\description{
Produce a tree graph from a hiearchical cluster analysis.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{ecdfPlot}.
}
\examples{
\dontrun{
# For an exmaple of a dendGram  see
vignette(topic="GraphGallery", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{probPlot}}
}
\keyword{hplot}

