% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfPlot.R
\name{ecdfPlot}
\alias{ecdfPlot}
\alias{ecdfPlot.default}
\alias{ecdfPlot.default}
\title{Empirical Distribution Plot}
\usage{
ecdfPlot(x, group = NULL, Plot = list(name = "Auto", what = "stairstep",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "Auto"), xaxis.log = TRUE, xaxis.range = c(NA, NA),
  xlabels = 11, ylabels = 5, ytitle = "Cumulative Probability",
  xtitle = deparse(substitute(x)), caption = "", margin = c(NA, NA, NA,
  NA), ...)

\method{ecdfPlot}{default}(x, group = NULL, Plot = list(name = "Auto", what
  = "stairstep", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "Auto"), xaxis.log = TRUE, xaxis.range = c(NA,
  NA), xlabels = 11, ylabels = 5, ytitle = "Cumulative Probability",
  xtitle = deparse(substitute(x)), caption = "", margin = c(NA, NA, NA,
  NA), ...)
}
\arguments{
\item{x}{the data to plot.}

\item{group}{create groups for \code{x}. Each group is plotted as a separate
line.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{xaxis.log}{logical, if \code{TRUE}, then log-transform the x axis.}

\item{xaxis.range}{set the range of the x-axis. See \bold{Details}.}

\item{xlabels}{set the x-axis labels. See \code{\link{linearPretty}} for
details.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}

\item{\dots}{any additional arguments needed by specific methods.}
}
\value{
Information about the graph.
}
\description{
Creates a graph of the empirical distribution function of data.
}
\details{
For linear axes, the range can be set to virtually any pair of values. For 
log axes, the choice of range is more resticted---for less than one log-cycle,
powers of whole numbers can be used; from 1 to about 3 log cycles, the choces 
should be powers of 3 or 10; and for more than 3 log cycles, the range sould be 
expressed only in powers of 10.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{ecdfPlot}.
}
\examples{
\dontrun{
set.seed(1)
X <- rlnorm(32)
setGD()
ecdfPlot(X)
# For more details of ecdfPlot see
vignette(topic="ProbabilityPlots", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{probPlot}}
}
\keyword{hplot}
