% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logPretty}
\alias{logPretty}
\title{Pretty Axis}
\usage{
logPretty(x, hard = FALSE, labels = "Auto", style = "numeric",
  extend.pct = 0, extend.range = NA)
}
\arguments{
\item{x}{data defining the range to be plotted on the axis. Missing value
are permitted, but ignored.}

\item{hard}{use the minimum and maximum of \code{x} as the fixed range of
the axis?}

\item{labels}{either "Auto," which lets the function decide how many labels,
the approximate number of labels, or the actual labels to use.}

\item{style}{a character string indicating the style of the axis labels if
they are not specifically listed in \code{labels}.}

\item{extend.pct}{extend the axis range by \code{extend.pct}. Only valid
when \code{hard} is FALSE.}

\item{extend.range}{if \code{TRUE}, then extend the data range by a bit to
avoid plotting on axis. Otherwise do not extend the data range. Only valid
when \code{hard} is FALSE; Ignored in \code{logPretty}}
}
\value{
Information about the axis lables.
}
\description{
Construct information for making a nicely formatted log-scale numeric axis.
}
\seealso{
\code{\link{areaPlot}}, \code{\link{boxPlot}},
\code{\link{colorPlot}}, \code{\link{areaPlot}}, \code{\link{dotPlot}},
\code{\link{ecdfPlot}}, \code{\link{probPlot}}, \code{\link{qqPlot}},
\code{\link{scalePlot}}, \code{\link{splomPlot}}, \code{\link{timePlot}},
\code{\link{xyPlot}}
}
\keyword{dplot}

