% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{month.USGS}
\alias{month.USGS}
\title{Month Abbreviations}
\format{A named character vector of the 12 preferred forms for month abbreviations.}
\usage{
month.USGS
}
\description{
A vector of USGS-style month abbreviations.
}
\examples{
\dontrun{
# See for examples of month.USGS:
vignette(topic="LineScatter", package="smwrGraphs")
}
}
\references{
Hansen, W.R., 1991, Suggestions to Authors of the United States Geological Survey,
7th Ed.: U.S. Govenement Printing Office, 289 p.
}
\keyword{datasets}

