% Generated by roxygen2 (4.0.2): do not edit by hand
\name{namePretty}
\alias{namePretty}
\title{Pretty Axis}
\usage{
namePretty(x, orientation = "table", order = "none", label.abbr = FALSE,
  offset = 0.5)
}
\arguments{
\item{x}{the discrete data values.}

\item{orientation}{the orientation of the data in \code{x}.\cr \tabular{lr}{
"table" \tab first in sequence at top (ends on right if x-axis)\cr "grid"
\tab first in sequence at bottom \cr }}

\item{order}{the oorder of the data in \code{x}.\cr \tabular{lr}{ "none"
\tab accept order as is\cr "ascending" \tab sort in ascending alphabetical
order\cr "descending" \tab sort in descending alphabetical order\cr named
vector \tab sort by values (largest value at top if orientation is
"table")\cr }}

\item{label.abbr}{crate abbreviations for \code{x}?}

\item{offset}{amount to offset the range, generally 0.5 or 1. The range of
the data is from 1 to the number of elements in in \code{x}.}
}
\value{
Information about the axis labels
}
\description{
Construct information for making a nicely formatted axis for discrete data.
}
\seealso{
\code{\link{dotPlot}}
}
\keyword{dplot}

