\name{numericData}
\alias{numericData}

\title{
Numeric Vlues
}
\description{
Converts data to numeric values (support function).
}
\usage{
numericData(x, lev = NULL)
}
\arguments{
  \item{x}{
any vector that can be converted to numeric.
}
  \item{lev}{
levels for character data that represent categories rather than character representations of numeric values.
}
}
\value{
Numeric data represented by \code{x}.
}
\keyword{ dplot }
