% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preSurface}
\alias{preSurface}
\title{Prepare for surface plot}
\usage{
preSurface(x, y, z.surf, zaxis.log = FALSE, zaxis.range = c(NA, NA),
  yaxis.log = FALSE, yaxis.range = c(NA, NA), xaxis.log = FALSE,
  xaxis.range = c(NA, NA), xlabels = "Auto", ylabels = "Auto",
  zlabels = "Auto", phi = NA, theta = NA, batch = FALSE)
}
\arguments{
\item{x}{the x-axis coordinate data, must be strictyly increasing. Missing values
not permitted. May be of type Date.}

\item{y}{the y-axis coordinate data, must be strictyly increasing. Missing values
not permitted.}

\item{z.surf}{a numeric matrix representing the surface. The length of \code{x}
must match the number of rows in \code{z.surf}. The length of \code{y}
must match the number of columns in \code{z.surf}.}

\item{zaxis.log}{log-transform the z axis?}

\item{zaxis.range}{set the range of the z-axis.}

\item{yaxis.log}{log-transform the y axis?}

\item{yaxis.range}{set the range of the y-axis.}

\item{xaxis.log}{log-transform the x axis?}

\item{xaxis.range}{set the range of the x-axis.}

\item{xlabels}{set up x-axis labels.}

\item{ylabels}{set up y-axis labels.}

\item{zlabels}{set up z-axis labels.}

\item{phi}{the viewing angle relative to the x-y plane. If \code{NA}, then
programmatically select a reasonable angle. Should be greater than 0 and less
than 90 degrees, but the best angles are generally between 20 and 45 degrees.}

\item{theta}{the viewing angle relative to the x-axis. Positive values rotate
the x-y plane in a clock-wise direction. If \code{NA}, then programmatically
select a reasonable angle by putting the largest values of \code{z.surf} near
the rear and the smallest values near the front. In general, the best angles
are not multiples of 90 degrees.}

\item{batch}{logical or character. If logical and \code{TRUE}, then select the
viewing angles specified by \code{phi} and \code{theta}. If logical and \code{FALSE},
then draw 9 candidate combinations of viewing angles and pause for user input
to select the desired viewing angles. If character, then select the viewing
angles specified by the letter selection, must be A through I.}
}
\value{
A list containg the projection inforamtion and the data for plotting. Must
be used in the call to \code{surfacePlot}
}
\description{
Select the projection for a surface plot.
}
\seealso{
\code{\link{surfacePlot}}
}

