% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Layout.R
\name{print.Layout}
\alias{print.Layout}
\title{Print Objects}
\usage{
\method{print}{Layout}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{\dots}{not used, required for other methods}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Prints the layout of a figure.
}
\note{
The layout of graphs is displayed.
}
\seealso{
\code{\link{setLayout}}, \code{\link{setGraph}}
}

