% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqPlot.R
\name{qqPlot}
\alias{qqPlot}
\alias{qqPlot.default}
\title{Q-Q Plot}
\usage{
qqPlot(x, y, alpha = 0.4, Plot = list(name = "Paired data quantiles",
  what = "points", type = "solid", width = "standard", symbol = "circle",
  filled = TRUE, size = 0.09, color = "black"), LineRef = list(name =
  "Line of best fit", what = "lines", color = "black"),
  Line1.1 = list(name = "Line of equality", what = "lines", color =
  "gray"), yaxis.log = FALSE, yaxis.range = c(NA, NA),
  xaxis.log = FALSE, xaxis.range = c(NA, NA), ylabels = 7,
  xlabels = 7, xtitle, ytitle, caption = "", margin = c(NA, NA, NA,
  NA), ...)

\method{qqPlot}{default}(x, y, alpha = 0.4, Plot = list(name =
  "Paired data quantiles", what = "points", type = "solid", width =
  "standard", symbol = "circle", filled = TRUE, size = 0.09, color =
  "black"), LineRef = list(name = "Line of best fit", what = "lines",
  color = "black"), Line1.1 = list(name = "Line of equality", what =
  "lines", color = "gray"), yaxis.log = FALSE, yaxis.range = c(NA, NA),
  xaxis.log = FALSE, xaxis.range = c(NA, NA), ylabels = 7,
  xlabels = 7, xtitle, ytitle, caption = "", margin = c(NA, NA, NA,
  NA), ...)
}
\arguments{
\item{x}{the x-axis data, or data to plot if \code{y} is missing.}

\item{y}{the y-axis data. If missing, then produce a quantile-normal
quantile plot from the data in \code{x}.}

\item{alpha}{the alpha value of the function for computing plotting
positions.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{LineRef}{control parameters of the reference line (best fit between
\code{x} and \code{y}. See \bold{Details}.}

\item{Line1.1}{control parameters for the 1:1 line. Drawn only for q-q plot.
See \bold{Details}.}

\item{yaxis.log}{logical, if \code{TRUE}, then log-transform the y axis.}

\item{yaxis.range}{set the range of the y axis. See \bold{Details}.}

\item{xaxis.log}{logical, if \code{TRUE}, then log-transform the x axis.}

\item{xaxis.range}{set the range of the x-axis. See \bold{Details}.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set the x-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}

\item{\dots}{any additional arguments required for specific methods.}
}
\value{
Information about the graph.
}
\description{
Creates a quantile-quantile (q-q) or a q-normal plot.
}
\details{
The argument \code{what} for either \code{LineRef} or \code{Line1.1} may be
set to "none" to suppress drawing of either line.

For linear axes, the range can be set to virtually any pair of values. For 
log axes, the choice of range is more resticted---for less than one log-cycle,
powers of whole numbers can be used; from 1 to about 3 log cycles, the choces 
should be powers of 3 or 10; and for more than 3 log cycles, the range sould be 
expressed only in powers of 10.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{qqPlot}.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
setGD()
qqPlot(X)
# For more details of qqPlot see
vignette(topic="ProbabilityPlots", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{ecdfPlot}},
\code{\link{probPlot}}
}
\keyword{hplot}
