% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportGraph.R
\name{reportGraph}
\alias{reportGraph}
\title{Report Graph}
\usage{
reportGraph(x, family = "Auto", size = 60 * par("csi"))
}
\arguments{
\item{x}{any R object.}

\item{family}{the font family to use in the report. See \bold{Details}.}

\item{size}{the size of the text, in points}
}
\value{
In contrast to other high-level graphics functions in the \code{smwrGraphs}
package, this function returns nothing because nothing is expected to be added to
the graph and nothing contributed to a possible explanation.
}
\description{
Create a report of any R object in a graph.
}
\details{
The value for \code{family} can be any valid font family for the
device. In general "serif," "sans," "mono," and "USGS" are valid. The default,
"Auto" selects "USGS" for character vectors, and "mono" for any other object.
}
\note{
The report is always placed in the upper left hand corner of the graph and is
left justified. If the report is longer than the height of the graph or wider than
the width of the graph, then the report is truncated.
}
\examples{
\dontrun{
setGD()
reportGraph("Hello world!")
# For more details of reportGraph see
vignette(topic="GraphGallery", package="smwrGraphs")
}
}

