% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seasonPlot.R
\docType{methods}
\name{seasonPlot}
\alias{seasonPlot}
\alias{seasonPlot,ANY,numeric-method}
\alias{seasonPlot,character,numeric-method}
\title{Season Plot}
\usage{
seasonPlot(x, y, Plot = list(), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.range = "", ylabels = 7, xlabels = 7,
  xtitle = "", ytitle = "", caption = "", margin = c(NA, NA, NA, NA),
  ...)

\S4method{seasonPlot}{ANY,numeric}(x, y, Plot = list(name = "", what =
  "points", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.range = c("calendar", "water", "climate"),
  ylabels = 7, xlabels = "Auto", xtitle = "",
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), ...)

\S4method{seasonPlot}{character,numeric}(x, y, Plot = list(name = "", what =
  "lines", type = "solid", width = "standard", symbol = "circle", filled = TRUE,
  size = 0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.range = c("calendar", "water", "climate"),
  ylabels = 7, xlabels = "Auto", xtitle = "",
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), ...)
}
\arguments{
\item{x}{the x-coordinate data.}

\item{y}{the y-coordinate data.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{yaxis.log}{logical, if \code{TRUE}, then log-transform the y axis.}

\item{yaxis.rev}{logical, if \code{TRUE}, then reverse the y axis.}

\item{yaxis.range}{set the range of the y-axis.}

\item{xaxis.range}{set the range of the x-axis. Must be one of "calendar,"
"water," or "climate" to set the type of year that is shown on the x-axis.}

\item{ylabels}{set up y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set up x-axis labels. See \bold{Details}.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}

\item{...}{arguments for specific methods.}
}
\value{
Information about the graph.
}
\description{
Create a plot of data on a yearly cycle.
}
\details{
For \code{seasonPlot}, the value for \code{xlabels} must be one of "full," the full
month names; "abbrev," abbreviations; or "letter," the first letter of the
month. The default is "Auto," which will select an appropriate labeling scheme.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{seasonPlot}.

To add a plot to the graph created by \code{seasonPlot}, the x-axis data
must be expressed as decimal time relative to January 1. The function
\code{baseDay2decimal} can be used to convert data in the form of base day
to decimal time.
}
\section{Methods}{
 \describe{

\item{signature(x = "ANY", y = "numeric")}{ Create a seasonal
plot for any valid date data and numeric y data. }

\item{signature(x = "character", y = "numeric")}{ Create a
seasonal plot for date data in the form of month and day, like "Jan 01" or
"Jaunary 01." Typically used to plot daily mean values. } }
}
\examples{
\dontrun{
# the months function is in lubridate
X <- as.Date("2001-01-15") + months(0:11)
set.seed(1)
Y <- runif(12)
setGD()
seasonPlot(X, Y)
# For more details of seasonPlot see
vignette(topic="LineScatter", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{timePlot}},
\code{\link[smwrBase]{baseDay2decimal}}
}
\keyword{hplot}
\keyword{methods}

