% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seriesPlot}
\alias{seriesPlot}
\alias{seriesPlot.default}
\title{Series Plot}
\usage{
seriesPlot(x, SeasonLine = list(name = "", what = "vertical", color =
  "black"), SeasonPoint = list(name = "", what = "points", symbol = "circle",
  filled = TRUE, size = 0.09, color = "black"), yaxis.log = FALSE,
  yaxis.range = c(NA, NA), ylabels = 7, xlabels, xtitle = "",
  ytitle = "", caption = "", margin = c(NA, NA, NA, NA), ...)
}
\arguments{
\item{x}{data that can be treated as a regularly-spaced time series. Missing
values are permitted, but result in missing seasons.}

\item{SeasonLine}{control parameters of the lines in the plot. See
\bold{Details}.}

\item{SeasonPoint}{control parameters of the points in the plot. See
\bold{Details}.}

\item{yaxis.log}{log-transform the y axis?}

\item{yaxis.range}{set the range of the y axis.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set the x-axis labels and number of seasons when \code{x} is
a simple numeric vector, may be a single numeric value indicating the number
of seasons in \code{x} or a vector of the names of the seasons. See
\code{\link{namePretty}} for details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{the parameters of the margin of the plot area.}

\item{\dots}{any additional arguments required for specific methods.}
}
\value{
Information about the graph.
}
\description{
Produce a plot.
}
\details{
The argument \code{what} for \code{SeasonLine} must be either "lines" or
"vertical."  See \code{\link{monthplot}} for more information.\cr The
argument \code{what} for \code{SeasonPoint} can be set to "none" to suppress
drawing of that feature.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{seriesPlot}.
}
\seealso{
\code{\link{setPage}}, \code{\link{monthplot}},
\code{\link{seasonPlot}}
}
\keyword{hplot}

