% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setAxis.R
\name{setAxis}
\alias{setAxis}
\title{Set up an Axis}
\usage{
setAxis(data, axis.range, axis.log, axis.rev, axis.labels, ...)
}
\arguments{
\item{data}{the coordinates for the particular axis}

\item{axis.range}{set axis range.}

\item{axis.log}{logical, if \code{TRUE}, then log transform the axis.}

\item{axis.rev}{logical, if \code{TRUE}, then reverse the axis direction.}

\item{axis.labels}{set axis labels.}

\item{\dots}{additional arguments to the "pretty" functions.}
}
\value{
Information about the axis
}
\description{
Sets up axis information (support function).
}
\seealso{
\code{\link{linearPretty}}, \code{\link{logPretty}}
}
\keyword{dplot}

