\name{setColor}
\alias{setColor}
\title{
Colors
}
\description{
Checks or converts anything to valid colors (support function).
}
\usage{
setColor(Color)
}
\arguments{
  \item{Color}{
any kind of data that might be interpreted as a color.
}
}
\value{
The values in \code{Color} converted to a value that could be interpreted as a color.
}
\seealso{
\code{link{colors}}, \code{link{rainbow}}
}
\keyword{ dplot }
