\name{setDefaults}
\alias{setDefaults}
\title{
Default Values
}
\description{
Sets the default values for control arguments (support function).
}
\usage{
setDefaults(current = list(), ...)
}
\arguments{
  \item{current}{
the control parameters specified in the call to the high-level graphing function.
}
  \item{\dots}{
the default values for each name required for the control parameters.
}
}
\value{
The control parameters with defaults substituted for missing names.
}
\keyword{ dplot }
