% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setGD}
\alias{setGD}
\alias{setPDF}
\alias{setPage}
\alias{setSweave}
\title{Graphics Page}
\usage{
setGD(name = "USGS")

setPage(layout = "portrait", font = "preview", name = "USGS",
  multiple = FALSE, device = "default")

setPDF(layout = "portrait", basename = "USGS", multiplefiles = FALSE)

setSweave(name, width, height, ...)
}
\arguments{
\item{name}{the name of the graphics page or the filename for
\code{setSweave}.}

\item{layout}{A description of the orientation and shape of the graphics
page. See \bold{Details}.}

\item{font}{a description of the font. The choices are "preview," which is
12 point Arial Narrow; "USGS," which is 8 point Arial Narrow; "EST," which
is 8 point Times New Roman; "PPT," which is 24 point Arial; and "PDF," which
is 8 point Arial Helvetica-Narrow. "PDF" should be chosen if the graphs are
to be saved to a portable document format (pdf) file.}

\item{multiple}{allow multiple pages?}

\item{device}{the name of the graphics device. See \bold{Details}.}

\item{basename}{the base name of the pdf file name.}

\item{multiplefiles}{modify \code{basename} to create multiple files for
multiple pages?}

\item{width}{the widht of the graphics page.}

\item{height}{the height of the graphics page.}

\item{\dots}{additional arguments, which are ignored by \code{setSweave}.}
}
\value{
For \code{setPage} and \code{setPDF}, a list with two components:
dev, the device number; and name, the name or basename. For \code{setSweave}
nothing is returned.
}
\description{
Sets up a graphics page.
}
\details{
If \code{layout} is "portrait," then the page size is 8.5 by 11 inches and
the graph area is 7.25 by 9.5 inches.\cr If \code{layout} is "landscape,"
then the page size is 11 by 8.5 inches and the graph area is 9.5 by 7.25
inches.\cr If \code{layout} is "square," then the page size is 7 by 7 inches
and the graph area is 6.5 by 6.5 inches (\code{setPage} only).\cr If
\code{layout} is "slide," then the page size is 10 by 7.5 inches and the
graph area is 9.5 by 7.0 inches (\code{setPage} only).\cr \code{Layout} may
also be a tagged list, with components \code{width} and \code{height} giving
the width and height of the page, the width and height of the graph area is
0.5 inch less that the page, except for \code{setPDF} where it is 0.1 inch
less.\cr The user may specify a graphics device other than the default for
the system. This may be necessary when running under certain user
environments like RStudio (TM).
}
\note{
The function \code{setSweave} is an interface to be used when using
\code{Sweave}. The functions \code{setSweave} and \code{setPDF} require a
call to \code{dev.off} to close the graphics device after all graphics is
completed.\cr The function \code{setGD} is designed to be a quick-and-dirty
graphics page set up function. It is designed to be used by functions to set
up the graphics environment if the user fails to do so.
}
\seealso{
\code{\link{setLayout}}, \code{\link{setGraph}}
}
\keyword{dplot}

