% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splomPlot.R
\name{splomPlot}
\alias{splomPlot}
\title{Scatter Plot Matrix}
\usage{
splomPlot(x, layout, Plot = list(name = "", what = "points", type = "solid",
  width = "standard", symbol = "circle", filled = TRUE, size = 0.05, color =
  "black"), Panel = list(), axis.log = FALSE, axis.range = c(NA, NA),
  labels = 5, caption = "")
}
\arguments{
\item{x}{the data to plot, must be either a matrix or a data frame.}

\item{layout}{the output from \code{setSplom}}

\item{Plot}{control parameters for each plot.}

\item{Panel}{a list or a function, see \bold{Details}.}

\item{axis.log}{log-transform the x and y axis for all plots.}

\item{axis.range}{set x- and y-axis ranges for all plots.}

\item{labels}{set the number of labels for both the x and y axes. See
\code{\link{linearPretty}} for details.}

\item{caption}{the figure caption.}
}
\value{
Information about the graph.
}
\description{
Produce a matrix of scatter plots
}
\details{
\code{Panel} may be a tagged list, with any one of these options:\cr
loess=span, where span is the span argument to loess.smooth;\cr line=opt,
where opt='slr' for simple linear regression, or 'loc' for line of organic
correlation, or '1:1' for the 1:1 line.\cr The format of the lines is taken
from \code{Plot}.

\code{Panel} may also be a function with 3
arguments, x, y, and current, that adds to the plot and returns updated plot
information. The function is called for each individual plot.
}
\note{
A call must be made to \code{setPage} and to \code{setSplom} to set up
the graphics environment before calling \code{splomPlot}.
}
\examples{
\dontrun{
# See for examples of splomPlot:
vignette(topic="LineScatter", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{setSplom}}
}
\keyword{hplot}

