% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stiffPlot.R
\name{stiffPlot}
\alias{addStiff}
\alias{stiffPlot}
\title{Stiff Diagrams}
\usage{
stiffPlot(cations, anions, Stiff = list(fill = "gray50", outline = "black",
  height = 2/3, bar = "black"), yaxis.orient = "table",
  yaxis.order = "none", xaxis.range = c(NA, NA), ylabels = "Auto",
  xlabels = 7, catlabels = "Auto", anlabels = "Auto",
  xtitle = "Milliequivalents per liter", ytitle = "", caption = "",
  margin = c(NA, NA, NA, NA), ...)

addStiff(x, y, width, height, cations, anions, Stiff = list(fill = "gray50",
  outline = "black", height = 2/3, bar = "black"), xaxis.range = c(NA, NA),
  catlabels = "", anlabels = "", current = list(yaxis.log = FALSE,
  yaxis.rev = FALSE, xaxis.log = FALSE))
}
\arguments{
\item{cations}{a matrix of cation data. Each row corresponds to the
respective \code{x} and \code{y} value. Missing values are not permitted for
\code{addStiff}, but are permitted for \code{stiffPlot} and result in no
Stiff Diagram for that entry.}

\item{anions}{a matrix of anion data. Each row corresponds to the respective
\code{x} and \code{y} value. Missing values are not permitted for
\code{addStiff}, but are permitted for \code{stiffPlot} and result in no
Stiff Diagram for that entry.}

\item{Stiff}{a list describing the Stiff diagram. See \bold{Details}.}

\item{yaxis.orient}{orientation of the y-axis values, must be either "table"
or "grid." "Table" is sorted from top to bottom, "grid" is sorted from
bottom to top.}

\item{yaxis.order}{the order of the y-axis values, must be one of "none,"
"ascending," or "descending."}

\item{xaxis.range}{the range of the x-axis corresponding to \code{width} in
the call to \code{addStiff} or the range of the x-axis in the call to
\code{stiffPlot}. See \bold{Details}.}

\item{ylabels}{set up y-axis labels.}

\item{xlabels}{set up x-axis labels.}

\item{catlabels}{labels for the values of the cations. For \code{addStiff},
the labels are applied to each Stiff diagram and for \code{stiffPlot}, the
labels are stored and drawn on the explanation \code{addExplanation}.}

\item{anlabels}{labels for the values of the anions. For \code{addStiff}, the
labels are applied to each Stiff diagram and for \code{stiffPlot}, the
labels are stored and drawn on the explanation \code{addExplanation}.}

\item{xtitle}{x-axis title (also called x-axis caption).}

\item{ytitle}{y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}

\item{x}{the x-coordinates to place the center of each Stiff diagram.
Missing values are permitted, but result in no Stiff diagram.}

\item{y}{the y-coordinates to place the center of each Stiff diagram.
Missing values are permitted, but result in no Stiff diagram.}

\item{width}{the width in inches of the Stiff diagrams.}

\item{height}{the height in inches of the Stiff diagrams.}

\item{current}{the current plotting parameters. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}

\item{\dots}{not used, required for other methods.}
}
\value{
Information about the graph.
}
\description{
Add a Stiff diagram to an existing graph or create a tabular presentation of
Stiff diagrams in a graph.
}
\details{
The units of the cation and anion data are generally in milli-equivalents
per liter.

The \code{Stiff} argument must be a tagged list with these components:
\describe{ \item{fill}{the name of the color to fill each Stiff diagram.
Must be a valid color name.} \item{outline}{the name of the color to draw
the outline or border for each Stiff diagram.  Must be a valid color name.}
\item{height}{the height of each Stiff diagram, proportional to the overall
height for each Stiff diagram.} \item{bar}{the color of the central bar. May
be "none" for no central bar.} }

The values for \code{axis.range} must be expressed as a negative value for the
\code{cation} data, left-hand side of the diagram, and a positive value for
\code{anion} data, right-hand side of the diagram.
}
\examples{
\dontrun{
# See for examples of stiffPlot:
vignette(topic="PiperPlot", package="smwrGraphs")
# See for examples of addStiff:
vignette(topic="GraphGallery", package="smwrGraphs")
}
}
\references{
Hem J.D., 1989, Study and interpretation of the chemical
characteristics of natural water: U.S. Geological Survey Water-Supply Paper
2254, 263 p.
}
\seealso{
\code{\link{addStiff}}, \code{\link{xyPlot}}
}
\keyword{hplot}

