% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryPlot.R
\name{ternaryPlot}
\alias{ternaryPlot}
\title{Ternary Diagram}
\usage{
ternaryPlot(x, y, z, Plot = list(name = "", what = "points", type = "solid",
  width = "standard", symbol = "circle", filled = TRUE, size = 0.09, color =
  "black"), axis.range = c(0, 100), num.labels = 6, ticks = TRUE,
  grids = !ticks, orient = "c", xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), ztitle = deparse(substitute(z)),
  units.title = "Percent", caption = "", margin = c(NA, NA, NA, NA))
}
\arguments{
\item{x}{the x-axis (bottom) data.}

\item{y}{the y-axis (left side) data.}

\item{z}{the z-axis (right side) data. Note that \code{x}, \code{y}, and
\code{z} do not need to sum to the axis range.}

\item{Plot}{control parameters of the plot, see \code{link{setMultiPlot}}
and \bold{Details} for details.}

\item{axis.range}{the range of the axes. Must be either c(0, 1) or c(0,
100).}

\item{num.labels}{the number of labels to draw on each axis. Best selections
are 2 giving (0, 100), 3 (0, 50, 100), 5 (0, 25, 50, 75, 100), or 6 (o, 20,
40, 60, 80, 100).}

\item{ticks}{logical, if \code{TRUE}, then draw ticks.}

\item{grids}{logical, if \code{TRUE}, then draw grid lines.}

\item{orient}{the orientation of the graph. Must be "c" for clockwise or "a"
for anti- or counter-clockwise}

\item{xtitle}{title (also called caption) for the x-axis.}

\item{ytitle}{title (also called caption) for the y-axis.}

\item{ztitle}{title (also called caption) for the z-axis.}

\item{units.title}{the units titles, should be either "Percent" of
"Proportion" depending on \code{axis.range}.}

\item{caption}{the figure caption.}

\item{margin}{set up the plot area margins--- ignored, included for
consistency with other plotting functions in this package.}
}
\value{
Information about the graph.
}
\description{
Produce a ternary diagram, also called a trilinear or triangular diagram.
}
\details{
The \code{what} component of the \code{Plot} argument must be either
"points" or "none."
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{piperPlot}.
}
\examples{
\dontrun{
# See for examples of ternaryPlot:
vignette(topic="PiperPlot", package="smwrGraphs")
}
}
\references{
Lorenz D.L., 2015, smwrGraphs
}
\seealso{
\code{\link{setPage}}, \code{\link{setMultiPlot}},
\code{\link{piperPlot}}, \code{\link{addTernary}}
}
\keyword{hplot}

