\name{timePretty}
\alias{timePretty}

\title{
Pretty Axis 
}
\description{
Construct information for making a "pretty" date axis.
}
\usage{
timePretty(x, labels = "Auto")
}

\arguments{
  \item{x}{
time difference data
}
  \item{labels}{
either "Auto," which lets the function decide how many labels, the approximate number of labels,
or the actual labels to use.
}
}
\value{
information about the axis lables.
}
\seealso{
\code{\link{timePlot}}
}
\keyword{ dplot }
