% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transPretty.R
\name{transPretty}
\alias{transPretty}
\title{Pretty Axis}
\usage{
transPretty(x, hard = FALSE, labels = 11, style = "none", func = log,
  Ifunc = exp, ...)
}
\arguments{
\item{x}{data defining the range to be plotted on the axis. Missing value
are permitted, but ignored.}

\item{hard}{logical, if \code{TRUE}, then use the minimum and maximum of
\code{x} as the fixed range of the axis, otherwise find "nice" limits.}

\item{labels}{either "Auto," which lets the function decide how many labels,
the approximate number of labels, or the actual labels to use.}

\item{style}{a character string indicating the style of the axis labels if
they are not specifically listed in \code{labels}.}

\item{func}{the forward transform function.}

\item{Ifunc}{the backward (inverse) transform function.}

\item{\dots}{additional arguments to \code{func} and \code{Ifunc}.}
}
\value{
Information about the axis lables.
}
\description{
Constructs information for making a nicely formatted numeric axis.
}
\seealso{
\code{\link{transPlot}}
}
\keyword{dplot}

