\name{USGSadd}
\alias{addXY}
\title{
Add a Plot
}
\description{
Adds a line or scatter plot to a graph.
}
\usage{
addXY(x, y, ...)
}
\arguments{
  \item{x}{
the x-coordinate data.
}
  \item{y}{
the y-coordinate data.
}
  \item{\dots}{
other arguments for specific methods.
}
}
\value{
Information about the plot.
}

\seealso{
\code{\link{addXY-methods}}
}
\keyword{aplot}

