% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{xyPlot}
\alias{xyPlot}
\alias{xyPlot,factor,numeric-method}
\alias{xyPlot,numeric,numeric-method}
\title{Plot Data}
\usage{
xyPlot(x, y, Plot = list(), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = 7, xtitle = "", ytitle = "", caption = "",
  margin = c(NA, NA, NA, NA), ...)

\S4method{xyPlot}{numeric,numeric}(x, y, Plot = list(name = "", what =
  "points", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = 7, xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), ...)

\S4method{xyPlot}{factor,numeric}(x, y, Plot = list(name = "", what =
  "points", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = "Auto", xtitle = "",
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), xlabels.rotate = FALSE, ...)
}
\arguments{
\item{x}{the x-axis data}

\item{y}{the y-axis data to plot}

\item{Plot}{control parameters of the plot.}

\item{yaxis.log}{log-transform the y axis?}

\item{yaxis.rev}{reverse the y axis?}

\item{yaxis.range}{set the range of the y-axis.}

\item{xaxis.log}{log-transform the x axis?}

\item{xaxis.range}{set the range of the x-axis.}

\item{ylabels}{set up y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set up x-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set up the plot area margins.}

\item{...}{arguments for specific methods.}

\item{xlabels.rotate}{rotate x-axis labels 90 degrees (perpendicular to the axis)?}
}
\value{
Information about the graph
}
\description{
Create a line or scatter plot.
}
\details{
Setting \code{ylabels} or \code{xlabels} to 0 will suppress ticks and
labels. That can be useful for relative axes or specialized labeling.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{xyPlot}.
}
\section{Methods}{
 \describe{
\item{signature(x = "numeric", y =
"numeric")}{ Create a line or scatter plot from numeric x and
y data. }
\item{signature(x = "factor", y ="numeric")")}{ Create a vertical
dot plot. Also useful for setting up a bar chart for discrete x-axis values. }
}
}
\seealso{
\code{\link{setPage}}, \code{\link{timePlot}},
\code{\link{colorPlot}}
}
\keyword{hplot}
\keyword{methods}

