% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addAnnotation.R
\name{addAnnotation}
\alias{addAnnotation}
\title{Add Text to a Graph}
\usage{
addAnnotation(x, y, annotation, leaderx = NULL, leadery = NULL,
  leadercol = "black", angle = 0, justification = "left", size = 60 *
  par("csi"), position = "above", current = list(yaxis.log = FALSE,
  yaxis.rev = FALSE, xaxis.log = FALSE))
}
\arguments{
\item{x}{the x-axis placement of \code{annotation}.}

\item{y}{the y-axis placement of \code{annotation}.}

\item{annotation}{the text.}

\item{leaderx}{draw leader from \code{x} to \code{leaderx}.}

\item{leadery}{draw leader from \code{y} to \code{leadery}.}

\item{leadercol}{the color of the leader.}

\item{angle}{the angle to rotate the text.}

\item{justification}{the justification of the text relative to \code{x},
\code{y}. Must be one of "left," "center," or "right."}

\item{size}{size of the text in points, the default is the current point size.}

\item{position}{the vertical location of the text. Must be one of "above,"
"below," or "center."}

\item{current}{the current plot controls. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}
}
\value{
The current plot information is returned invisibly.
}
\description{
Adds text to a Graph.
}
\details{
This function places only a single text on the graph for each call.\cr A
leader from \code{x}, \code{y} to \code{leaderx}, \code{leadery} if
\code{leaderx} is not \code{NULL}.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
AA.pl <- xyPlot(X, Y)
# Add the median line of Y and annotation
refLine(horizontal = median(Y), current=AA.pl)
addAnnotation(min(X), median(Y), "Median Y", current=AA.pl)
# For more details of addAnnotation see
vignette(topic="GraphAdditions", package="smwrGraphs")
demo(topic="TopAxisExample", package="smwrGraphs")
}
}
\seealso{
\code{\link{labelPoints}}, \code{\link{addTable}},
\code{\link{xyPlot}}
}
\keyword{aplot}

