% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addAxisLabels.R
\name{addAxisLabels}
\alias{addAxisLabels}
\title{Axis Ticks and Labels}
\usage{
addAxisLabels(which, current, title = "", ticks = FALSE, labels = TRUE)
}
\arguments{
\item{which}{which axis to label, must be one of "bottom," "left," "top," or
"right."}

\item{current}{a list containing the current plot information, see \bold{Details}.}

\item{title}{the axis title.}

\item{ticks}{draw the ticks.}

\item{labels}{draw the labels.}
}
\value{
The current plot information is returned invisibly.
}
\description{
Add axis ticks, labels, and title
}
\details{
The \code{current} argument is generally the output from a high-level
plotting function in \code{smwrGraphs}. If \code{which} is "left" or "right,"
then \code{current} must contain a component named \code{yax}. If  \code{which} is
"bottom" or "top," then \code{current} must containa component named \code{xax}.
Those components are generally constructed from functions like \code{linearPretty} or
or \code{logPretty}.
}
\note{
In general, all functions that create plots will draw the necessary
axes. This function should be used only to add axis labels to an unlabeled
axis. Axis labels can be suppressed by setting up the margins with negative
values.
}
\examples{
\dontrun{
set.seed(1)
X <- runif(25)
Y <- runif(25)
AA.pl <- xyPlot(X, Y)
addAxisLabels("top", AA.pl, labels=TRUE)
# For more details of addAxisLabels see
vignette(topic="GraphSetup", package="smwrGraphs")
}
}
\seealso{
\code{\link{linearPretty}}, \code{\link{logPretty}},
\code{\link{datePretty}}, \code{\link{transPretty}}, \code{\link{addLabel}}
}
\keyword{aplot}

