% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBars.R
\name{addBars}
\alias{addBars}
\title{Bar Graph}
\usage{
addBars(x, y, base = 0, Bars = list(name = "Auto", fill = "gray80",
  outline = "black", width = "Auto", orientation = "stack"),
  current = list(yaxis.log = FALSE, yaxis.rev = FALSE, xaxis.log =
  FALSE))
}
\arguments{
\item{x}{the x-coordinate data. Missing values are permitted but result in
no bar.}

\item{y}{the heights of the bars or y-coordinate data. Missing values are
permitted but result in no bar. For stacked or grouped bars, \code{y} must be
a matrix, each row corresponding to the value in \code{x} and the columns
representing each bar. See \bold{Note}.}

\item{base}{extend the bars from this value.}

\item{Bars}{parameters defining the characteristics of the bars. See
\bold{Details}.}

\item{current}{the current plot information. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}. 
See \bold{Note}.}
}
\value{
The current plot information is returned invisibly.
}
\description{
Creates a bar chart by adding bars to an existing graph.
}
\details{
The \code{Bars} argument must be a tagged list with these components:
\describe{ 
\item{name}{a character vector describing each column of data; used in the 
explanation. If "Auto," then derive the name from the column names in \code{y}.}
\item{fill}{the name of the color to fill each bar. For multiple bars, can be
a vector of colors or the name of a color sequence generating function, such as
"pastel.colors."} 
\item{outline}{the name of the color to draw the outline or border for each 
bar. If "none," then no border is drawn.} 
\item{width}{the width of each bar in x-axis units. For discrete x-axis,
If \code{width} 1, then the bars form a continuous filled area.
The default is "Auto," which fills 2/3 of the distance. If 0, then draw vertical
lines rather than bars; the color of the line is based on \code{outline}.}
\item{orientation}{the orientation of the bars. Must be either "stack" or
"group." Can be abbreviated to a single letter.} }
}
\note{
Use of \code{addBars} adds 1 step to creating bar charts, but adds
flexibility in axis formatted from existing high-level plotting functions such
as \code{xyPlot} or \code{timePlot}.

Bars are only valid for linear y-axes. Calling \code{addBars} when 
\code{yaxis.log} or \code{yaxis.rev} is \code{TRUE} or for any arbitrary 
transform of the y-axis will cause \code{addBars} to fail.

Datasets containing grouped data are often stacked with a column indicating the grouping.
There are several functions that will reformat stacked datasets. The \code{group2row}
function is very flexible in accepting many types of data to reformat rather than
only numeric data.
}
\examples{
\dontrun{
set.seed(1)
X <- seq(1, 9, by=1.0)
Y <- runif(9) + runif(9)
setGD()
AA.pl <- xyPlot(X, Y, Plot=list(what="none"), yaxis.range=c(0,2))
addBars(X, Y, base=0, current=AA.pl)
# For more details of addBars see
demo(topic="AnnualFlowBarChart", package="smwrGraphs")
}
}
\seealso{
\code{\link{xyPlot}}, \code{\link{timePlot}}, \code{\link{addXY}}
\code{\link[smwrBase]{group2row}}
}
\keyword{aplot}
