% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addCaption}
\alias{addCaption}
\title{Add Caption}
\usage{
addCaption(caption = "")
}
\arguments{
\item{caption}{the text of the caption for the graph}
}
\value{
Nothing is returned.
}
\description{
Adds a caption at the bottom of the graph
}
\note{
Useful for adding 1-line captions.
}
\keyword{aplot}

