% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addCI}
\alias{addCI}
\title{Add Confidence Interval Lines}
\usage{
addCI(type, level = 0.95, Plot = list(name = "", what = "lines", type =
  "solid", width = "standard", color = "black"), current = list(yaxis.log =
  FALSE, yaxis.rev = FALSE, xaxis.log = FALSE))
}
\arguments{
\item{type}{the type of confidence interval desired. Must be either "SLR" for the confidence
interval for a simple linear regression model, which must have been created using the
\code{addSLR} function or "q-norm" for the confidence interval for a q-normal plot created
using \code{qqPlot}.}

\item{level}{the confidence level desired.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{current}{the current plot information. This would is the
output from one of the graph creation or addition functions like \code{qqPlot}
or \code{addSLR}.}
}
\value{
The current plot information, the x and y components are the data, not the line.
The regression model is included as the lm component.
}
\description{
Computes the confidence interval for a linear regression or q-normal graph and adds the lines to a graph.
}
\note{
The equation for the confidence intervals for a simple linear regression
model can be found in any textbook on regression, see section 9.4.4 in Helsel and
Hirsch (2002) for example.\cr
The confidence interval for a normal distribution is described in Appendix 9 in
U.S. Water Resources Council (1982). Owen (1968) describes the application of the
noncentral \emph{t}-distribution for computing the tolerance limits for a normal
distribition.
}
\examples{
\dontrun{
# See for examples of addCI:
vignette(topic="GraphAdditions", package="smwrGraphs")
}
}
\references{
Helsel, D.R., and Hirsch, R.M., 2002, Statistical methods in water resources:
U.S. Geological Survey Techniques of Water-Resources Investigations, book 4,
chap. A3, 522 p.\cr
U.S. Water Resources Council, 1982, Guidelines for determining flood flow frequency,
revised September 1981, Editorial Correciotns March 1982:
Hydrology Committee Bulletin 17B, Washington D.C., 190 p., 1 plate.
}
\seealso{
\code{\link{addSLR}}, \code{\link{qqPlot}}, \code{\link{addErrorBars}}
}
\keyword{aplot}

