% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addExplanation}
\alias{addExplanation}
\title{Add Explanation}
\usage{
addExplanation(what, where = "new", title = expression(bold(EXPLANATION)),
  box.off = where != "new", margin = rep(0, 4))
}
\arguments{
\item{what}{a specialized object for an explanation, from the output from
calls to the plotting functions}

\item{where}{a description of where to put the explanation, see
\bold{Details}.}

\item{title}{the title of the explanation.}

\item{box.off}{logical, if \code{TRUE}, then box off the explanation with a
blank background and black box, otherwise insert as is.}

\item{margin}{the margin for a new graph}
}
\value{
Nothing is returned.
}
\description{
Adds an explanation to the current graph or to a new graph.
}
\details{
\code{where} must be one of "ul," "ur," "ll," "lr," "cl," "cr," "uc," "lc,"
"cc," or "new." If "new," then the explanation is placed in a new graph,
otherwise, the first letter is an abbreviation for upper, lower, or center
and the second letter is an abbreviation for left, right, or center. the
explanation for a box plot must be placed in a new graph.
}
\note{
Box plot explanations require fairly large graph areas because of the
detail required for some types. In general, a graph about 4.5 inches high is
needed for the Tukey type and 4 inches for other types and widths of 2.5 and
2 inches respectively. If the graph area is smaller than required, then a
warning is printed and the explanation may be unreadable.
}
\keyword{aplot}

