% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addLabel.R
\name{addLabel}
\alias{addLabel}
\title{Axis Labels}
\usage{
addLabel(label, x, side = "bottom", size = "Auto", distance = 0.2,
  justification = "center", orientation = "parallel",
  current = list(yaxis.log = FALSE, yaxis.rev = FALSE, xaxis.log = FALSE))
}
\arguments{
\item{label}{the text or expression to add to the graph.}

\item{x}{the axis location in the correct user units.}

\item{side}{the side to place \code{label}. Must be "bottom," "left," "top,"
or "right." Only the first letter is necessary.}

\item{size}{the size of the text in points.}

\item{distance}{the distance from the axis, in lines of text.}

\item{justification}{defines the placement of the text relative to \code{x}
if \code{orientation} is "parallel" and relative to \code{distance} if
\code{orientation} is "perpendicular." Must be one of "left," "center," or
"right."}

\item{orientation}{the orientation of the label relative to the axis. Must
be either "parallel" or "perpendicular."}

\item{current}{the current plot parameters. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}
}
\value{
Nothing is returned.
}
\description{
Adds text in the margin for axis labels.
}
\note{
In general, all functions that create plots will draw the necessary
axes. This function should be used only to add axis labels to an unlabeled
axis. Axis labels can be suppressed by setting up the margins with negative
values or, for some functions, special arguments to \code{xlabels} or
\code{ylabels}.
}
\examples{
\dontrun{
# See for examples of addLabel:
vignette(topic="GraphGallery", package="smwrGraphs")
demo(topic="AnnualFlowBarChart", package="smwrGraphs")
}
}
\seealso{
\code{\link{mtext}} and \code{\link{plotmath}} for example
expressions, \code{\link{addAxisLabels}}
}
\keyword{aplot}

