% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addTitle.R
\name{addTitle}
\alias{addTitle}
\title{Add a Title}
\usage{
addTitle(Main = "", Heading = "", Justification = "left", Bold = TRUE,
  Position = "above")
}
\arguments{
\item{Main}{the main text of the title. Can be either a character string or an
"expression" object.}

\item{Heading}{The title heading, generally a single letter. See
\bold{Details}}

\item{Justification}{specify the horizontal location of the title, must be one of
"left," "center," or "right."}

\item{Bold}{logical, if \code{TRUE}, then display the title in bold face type.}

\item{Position}{specify the vertical location of the title, must be either
"above" or "inside."}
}
\value{
Nothing is returned.
}
\description{
Adds a title (also called heading) to a graph.
}
\details{
If only \code{Heading} is non blank, then the title is a single letter in
bold italics. If both \code{Heading} and \code{Main} are non blank, then the
title is a single letter followed by a period in bold italics followed by
\code{Main} in bold if \code{Bold} is \code{TRUE}.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
setGD()
AA.pl <- xyPlot(X, Y)
addTitle("X and Y")
# For more details of addTitle see
vignette(topic="BoxPlots", package="smwrGraphs")
vignette(topic="GraphGallery", package="smwrGraphs")
vignette(topic="GraphSetup", package="smwrGraphs")
}
}
\seealso{
\code{\link{addCaption}}, \code{\link{addAnnotation}},
\code{\link{addTable}}
}
\keyword{aplot}

