% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{addXY}
\alias{addXY}
\alias{addXY,ANY,numeric-method}
\alias{addXY,numeric,character-method}
\title{Add a plot to a graph}
\usage{
addXY(x, y, ...)

\S4method{addXY}{ANY,numeric}(x, y, Plot = list(name = "", what = "lines",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "black"), current = list(yaxis.log = FALSE, yaxis.rev = FALSE,
  xaxis.log = FALSE), right = FALSE, right.log = FALSE, right.rev = FALSE,
  right.range = c(NA, NA), right.labels = 7, right.title = "")

\S4method{addXY}{numeric,character}(x, y, Plot = list(name = "", what =
  "points", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "black"), current = list(yaxis.log = FALSE,
  yaxis.rev = FALSE, xaxis.log = FALSE), jitter.y = FALSE)
}
\arguments{
\item{x}{the x-axis data. Missing values are permitted, but result in breaks
in the plotted data.}

\item{y}{the y-axis data. Missing values are permitted, but result in breaks
in the plotted data.}

\item{...}{arguments for specific methods.}

\item{Plot}{parameters of the plot.  See \code{\link{setPlot}} for a
description of the parameters.}

\item{current}{the current plot information. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}

\item{right}{logical: set up new right axis?}

\item{right.log}{logical: log transform right axis?}

\item{right.rev}{logical: reverse right axis?}

\item{right.range}{set right-axis range.}

\item{right.labels}{set up right-axis labels.}

\item{right.title}{the right-axis title.}

\item{jitter.y}{adjust \code{y} values to reduce overlap for each group?}
}
\value{
Information about the graph.
}
\description{
Add points or lines to the current graph.
}
\section{Methods}{
 \describe{

\item{signature(x ="ANY", y = "numeric"}{Any valid x-axis data and numeric y.}
\item{signature(x ="numeric", y = "character"}{Method to add to a dot plot;
the right-axis arguments are not valid.}
}
}
\keyword{aplot}
\keyword{methods}

