% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/areaPlot.R
\name{areaPlot}
\alias{areaPlot}
\title{Shaded Area Plot}
\usage{
areaPlot(x, y, Areas = list(name = "Auto", fillDir = "between", base = "Auto",
  lineColor = "black", fillColors = "pastel"), yaxis.log = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = "Auto", xtitle = "", ytitle = "",
  caption = "", margin = c(NA, NA, NA, NA))
}
\arguments{
\item{x}{numeric x-axis coordinates in increasing order.}

\item{y}{a numeric matrix of y-axis coordinates.}

\item{Areas}{parameters controlling the areas. See \bold{Details}.}

\item{yaxis.log}{logical, if \code{TRUE}, then log transform y axis.}

\item{yaxis.range}{set y-axis range. See \bold{Details}.}

\item{xaxis.log}{logical, if \code{TRUE}, then log transform x axis.}

\item{xaxis.range}{set x-axis range. See \bold{Details}.}

\item{ylabels}{set up y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set up x-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the caption for the graph.}

\item{margin}{set up the plot area margins.}
}
\value{
Information about the graph.
}
\description{
Produces a plot where the area between lines is filled with color.
}
\details{
The components of \code{Areas} are \code{name}, the name or names
to use to describe the areas in the explanation, the default "Auto" generates
names from the column names of \code{y}; \code{fillDir}, how to
fill---must be either "between" or "under;" \code{base}, the base value when
\code{fillDir} is "under," can be "Auto" to draw to the x-axis or any numeric value;
\code{lineColor} specifies the color to draw the lines around each area, may be
"none" for no line drawn; \code{fillColors} specifies colors for each area, when only
a single area is drawn, then the value must be the name of a color, otherwise either a
vector of colornames or the prefix name of a function that generates a sequence of
colors. The prefix name is prepended to ".colors" for the name of the function. See the
documentation for these functions in the \bold{See Also} section.

For linear axes, the range can be set to virtually any pair of values. For
log axes, the choice of range is more resticted---for less than one log-cycle,
powers of whole numbers can be used; from 1 to about 3 log cycles, the choces
should be powers of 3 or 10; and for more than 3 log cycles, the range sould be
expressed only in powers of 10.
}
\examples{
\dontrun{
set.seed(1)
X <- seq(1, 9, by=.5)
Y <- runif(17) + runif(17)
# The default fillDir, between, requires at least a 2-column matrix
areaPlot(X, cbind(rep(0, 17),Y))
# For more details of areaPlot see
vignette(topic="GraphGallery", package="smwrGraphs")
}
}
\seealso{
\code{\link{addArea}}, \code{\link{smwr.colors}}, \code{\link[grDevices]{heat.colors}}
}

