% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biPlot.default.R
\name{biPlot.default}
\alias{biPlot.default}
\title{Biplot}
\usage{
\method{biPlot}{default}(x, y, separate.axes = TRUE, xPlot = list(name =
  "observations", what = "points", type = "solid", width = "standard", symbol =
  "circle", filled = TRUE, size = 0.05, color = "black"), yPlot = list(name =
  "variables", width = "color", size = 0.2, color = "darkblue", symbol =
  "arrow", filled = FALSE), xPlotLabels = list(labels = "rownames", dir =
  "NE", size = 8, offset = 0.35), yPlotLabels = list(labels = "colnames", dir
  = "Auto", size = 8, offset = 0.35, color = "darkblue"), ylabels = 5,
  xlabels = 5, ylabels2 = 5, xlabels2 = 5, xtitle = "", ytitle = "",
  xtitle2 = "", ytitle2 = "", range.factor = 1.25, caption = "",
  margin = c(NA, NA, NA, NA), ...)
}
\arguments{
\item{x}{a 2-column matrix of x- (column 1) and y- (column 2) coordinates for 
observations or equivalent.}

\item{y}{a 2-column matrix of x- (column 1) and y- (column 2) coordinates for
variables or equivalent.}

\item{separate.axes}{logical, if \code{TRUE}, then plot \code{x} and \code{y} data
on separate axes.}

\item{xPlot}{control information to plot the \code{x} data.  See
\code{\link{setPlot}} for a description of the parameters.}

\item{yPlot}{control information to plot the \code{y} data.  See
\code{\link{setPlot}} for a description of the parameters. For \code{yPlot},
symbol can be "arrow" to indicate that an arrow is to be drawn from the
origin.}

\item{xPlotLabels}{control information for the \code{x} data labels. See
\bold{Details}.}

\item{yPlotLabels}{control information for the \code{y} data labels. See
\bold{Details}.}

\item{ylabels}{set y-axis labels for \code{x} data. See
\code{\link{linearPretty}} for details.}

\item{xlabels}{set x-axis labels for \code{x} data. See
\code{\link{linearPretty}} for details.}

\item{ylabels2}{set y-axis labels for \code{y} data. See
\code{\link{linearPretty}} for details.}

\item{xlabels2}{set x-axis labels for \code{y} data. See
\code{\link{linearPretty}} for details.}

\item{xtitle}{x-axis title (also called x-axis caption) for \code{x} data.}

\item{ytitle}{y-axis title (also called y-axis caption) for \code{x} data.}

\item{xtitle2}{x-axis title (also called x-axis caption) for \code{y} data.}

\item{ytitle2}{y-axis title (also called y-axis caption) for \code{y} data.}

\item{range.factor}{a numeric factor by which to expand the axis ranges so
that labels can be drawn.}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}

\item{\dots}{not used, required for other methods.}
}
\value{
Information about the graph
}
\description{
Produces a biplot, which is a plot of two different types of data on the same
graph.
}
\details{
The \code{xPlotLabels} and \code{yPlotLabels} arguments must be tagged lists
with these components: \describe{ \item{labels}{the labels. For
\code{xPlotLabels}, "rownames" means use the row names from \code{x} to
generate the labels. For \code{yPlotLabels}, "colnames" means use the column
names from \code{y} to generate the labels. Otherwise a character vector of
the labels.} \item{dir}{the direction the label text is placed from the
object.} \item{size}{the size of the label text.} \item{offset}{the distance
the labels is placed relative to the object.} \item{color}{the color of the
label text.} }
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{biPlot}.
}
\seealso{
\code{\link{setPage}}, \code{\link{biPlot}}
}
\keyword{hplot}
