% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxPlotStats}
\alias{boxPlotStats}
\title{Compute Statistcs for a Box Plot}
\usage{
boxPlotStats(x, Box, yaxis.log)
}
\arguments{
\item{x}{a list containing the data to compute the statistics.}

\item{Box}{control paramters for the box.}

\item{yaxis.log}{logical: log transform data before computing the
statistics?}
}
\value{
a list containing the statistics for each box plot.
}
\description{
Compute the statistics for truncated, simple, Tukey, or extended box plots
(support function for \code{boxPlot}).
}
\seealso{
\code{\link{boxPlot}}
}
\keyword{dplot}

