% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlotStats.R
\name{boxPlotStats}
\alias{boxPlotStats}
\title{Compute Statistcs for a Box Plot}
\usage{
boxPlotStats(x, Box, yaxis.log)
}
\arguments{
\item{x}{a list containing the data to compute the statistics.}

\item{Box}{control parameters for the box as set by the user and verified by
\code{boxPlot}.}

\item{yaxis.log}{logical, if \code{TRUE}, then log transform data before computing the
statistics.}
}
\value{
a list containing the statistics for each box plot.
}
\description{
Computes the statistics for truncated, simple, Tukey, or extended box plots 
(support function for \code{boxPlot}).
}
\seealso{
\code{\link{boxPlot}}
}
\keyword{dplot}

