% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contourPlot.R
\name{contourPlot}
\alias{contourPlot}
\alias{contourPlot.default}
\alias{contourPlot.matrix}
\title{Contour Plot}
\usage{
contourPlot(z, ...)

\method{contourPlot}{default}(z, x, y, Grid = list(method = "interpolate",
  linear = TRUE, extrapolate = FALSE, density = 90, span = 0.75, degree = 1,
  family = "symmetric"), Contours = list(name = "Auto", levels = 10, filled =
  FALSE, lineColor = "black", lineLabel = "flattest", fillColors = "coolWarm"),
  yaxis.range = c(NA, NA), xaxis.range = c(NA, NA), ylabels = 4,
  xlabels = 4, xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), ...)

\method{contourPlot}{matrix}(z, rows, cols, matrix.rows = "x",
  Contours = list(name = "Auto", levels = 10, filled = FALSE, lineColor =
  "black", lineLabel = "flattest", fillColors = "coolWarm"),
  yaxis.range = c(NA, NA), xaxis.range = c(NA, NA), ylabels = 4,
  xlabels = 4, xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), ...)
}
\arguments{
\item{z}{the values representing the surface data.}

\item{x}{the x-axis coordinates for each value in \code{z}.}

\item{y}{the y-axis coordinates for each value in \code{z}.}

\item{Grid}{control paramters for gridding irregularly spaced data.}

\item{Contours}{control parameters for the coutour lines or levels in the filled plot.}

\item{yaxis.range}{set the range of the y-axis.}

\item{xaxis.range}{set the range of the x-axis.}

\item{ylabels}{set up y-axis labels. See \code{\link{linearPretty}} for details.}

\item{xlabels}{set up x-axis labels. See \code{\link{linearPretty}} for details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins.}

\item{rows}{the coordinates for \code{z} represented by the rows in the matrix.}

\item{cols}{the coordinates for \code{z} represented by the columns in the matrix.}

\item{matrix.rows}{a single character, either "x" or "y" indicating whether the rows in z should be
plotted along the x or y axis.}

\item{\dots}{not used, required for other methods.}
}
\value{
Information about the graph.
}
\description{
Produce a contour plot or a colored surface with colors corresponding to
values in \code{z}.
}
\details{
Missing values are permitted in \code{z}, \code{x}, and \code{y} for the default
method and are removed, with a warning. before constructing the surface.

Missing values are not permitted in \code{rows} or \code{columns} but are permitted
in \code{z} for the matrix method. Missing values in \code{z} result in blank areas
in the plot.
}
\examples{
\dontrun{
set.seed(1)
Xbig <- runif(100)
Ybig <- runif(100)
# Make a hill
Zbig <- 1 - ((Xbig-.5)^2 + (Ybig-.5)^2)^.75)
contourPlot(Zbig, Xbig, Ybig)
# See for examples of contourPlot see
vignette(topic="GraphGallery", package="smwrGraphs")
}
}

