% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/copyDemo.R
\name{copyDemo}
\alias{copyDemo}
\title{Copy a Demo File}
\usage{
copyDemo(topic, package = "smwrGraphs", file)
}
\arguments{
\item{topic}{the name of the topic, must be a character string.}

\item{package}{the package name, must be a character string.}

\item{file}{the target file name, must be a character string. If missing,
then the file name is created from the \code{topic} name.}
}
\value{
Either the name of the target file or \code{NULL} if the copy
failed.
}
\description{
Copies a demo file from the source package to a file.
}
\examples{
\dontrun{copyDemo("HydroPrecip")}
}
\seealso{
\code{\link{demo}}
}
\keyword{utilities}

