% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEllipse.R
\name{dataEllipse}
\alias{dataEllipse}
\title{Construct an Ellipse}
\usage{
dataEllipse(x, y, percent = 100, smooth = 0)
}
\arguments{
\item{x}{the x-coordinate data.}

\item{y}{the y-coordinate data.}

\item{percent}{a scale factor, adjusted to include \code{percent} of the
data.}

\item{smooth}{required for naming compatibility with other functions, not used.}
}
\value{
A list containing the x- and y-coordinates of the ellipse.
}
\description{
Constructs an ellipse from x- and y-coordinate data.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
TMP <- dataEllipse(X, Y)
# Just print the first 10 values
lapply(TMP, function(x) x[1:10])
# For examples of dataEllipse in graphs see:
vignette(topic="GraphAdditions", package="smwrGraphs")
}
}
\seealso{
\code{\link{cov2Ellipse}}, \code{\link{hull}}
}
\keyword{dplot}

