% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataEllipse}
\alias{dataEllipse}
\title{Construct an Ellipse}
\usage{
dataEllipse(x, y, percent = 100, smooth = 0)
}
\arguments{
\item{x}{the x-coordinate data.}

\item{y}{the y-coordinate data.}

\item{percent}{a scale factor, adjusted to include \code{percent} of the
data.}

\item{smooth}{required for compatibiltiy, not used.}
}
\value{
A list containing the x- and y-coordinates of the ellipse.
}
\description{
Construct an ellipse from x- and y-coordinate data.
}
\examples{
\dontrun{
# See for examples of dataEllipse:
vignette(topic="GraphAdditions", package="smwrGraphs")
}
}
\seealso{
\code{\link{cov2Ellipse}}, \code{\link{hull}}
}
\keyword{dplot}

