% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datePretty}
\alias{datePretty}
\title{Pretty Axis}
\usage{
datePretty(x, major = "Auto", minor = "Auto", tick.span = 1,
  style = c("Auto", "at", "between"), label.abbr = 0)
}
\arguments{
\item{x}{date and time data}

\item{major}{major tick interval, must be one of "hours," "days," "months,"
"years," or "Auto," which will make an intelligent choice from those. "Auto"
will also automatically change \code{tick.span}.}

\item{minor}{minor tick interval, must be one of "min," "hours," "days,"
"months," "years," or "Auto " which will make an intelligent choice from
those, possibly adjusted by a scaling factor. Can also be formatted as an
argument to \code{seq}, such as "15 mins" for 15-minute ticks.}

\item{tick.span}{span between major labels. For example, with "years"
option, tick.span=5 would generate labels like 1990, 1995, 2000 and so
forth.}

\item{style}{style of lables, must be one of "at," "between," or "Auto,"
which selects the "best" style. "At" places the labels at the ticks and
"between" places the labels between the major ticks.}

\item{label.abbr}{indicator of the degree of abbreviation for labels\cr
\tabular{lr}{ 0 \tab best guess based on number of intervals\cr 1 \tab
shortest (single letter month, for example)\cr 2 \tab standard abbreviation
(3-letter month name, for example)\cr 3 \tab full text\cr }}
}
\value{
information about the axis labels.
}
\description{
Construct information for making a "pretty" date axis.
}
\seealso{
\code{\link{timePlot}}
}
\keyword{dplot}

