% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendGram.R
\name{dendGram}
\alias{dendGram}
\title{Tree Graphs}
\usage{
dendGram(x, Tree = list(orientation = "vertical", width = "standard", color =
  "black"), axis.range = c(NA, NA), labels = "Auto", ytitle = "",
  xtitle = "", caption = "", margin = c(NA, NA, NA, NA))
}
\arguments{
\item{x}{the data to plot. Must be able to be converted to class 
"dendrogram."}

\item{Tree}{control parameters of the tree diagram. See \bold{Details}.}

\item{axis.range}{set the range of the tree-axis.}

\item{labels}{set the tree-axis labels. See \code{\link{linearPretty}} for
details.}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}
}
\value{
Information about the graph.
}
\description{
Produces a tree graph from a hierarchical cluster analysis.
}
\details{
\code{Tree} is a list with these components: 
\describe{ 
\item{orientation}{the orientation of the dendrogram, must be either "vertical" or
"horizontal"}
\item{width}{the line width of the line representing the dendrogram}
\item{color}{the color of the line representing the dendrogram}}
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{ecdfPlot}.
}
\examples{
\dontrun{
# For an example of a dendGram  see
vignette(topic="GraphGallery", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{probPlot}}
}
\keyword{hplot}
