% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/histGram.R
\name{histGram}
\alias{histGram}
\alias{histGram.default}
\title{Histogram}
\usage{
histGram(x, breaks = "Sturges", Hist = list(), yaxis.range = c(NA, NA),
  ylabels = 7, xlabels = 7, xtitle = "", ytitle = "Auto",
  caption = "", margin = c(NA, NA, NA, NA), ...)

\method{histGram}{default}(x, breaks = "Sturges", Hist = list(type =
  "frequency", fill = FALSE, boundary = "lower", line.color = "black",
  fill.color = "gray80"), yaxis.range = c(NA, NA), ylabels = 7,
  xlabels = 7, xtitle = deparse(substitute(x)), ytitle = "Auto",
  caption = "", margin = c(NA, NA, NA, NA), ...)
}
\arguments{
\item{x}{a numeric vector to create the histogram}

\item{breaks}{any valid value for \code{\link{hist}}. See \bold{Details}.}

\item{Hist}{Controls for the histogram.}

\item{yaxis.range}{set the range for the y axis, the first value must be 0.}

\item{ylabels}{the approximate number of labels for the y axis.}

\item{xlabels}{the approximate number of labels for the x axis.}

\item{xtitle}{x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption),
"Frequency" for a frequency histogram,
"Density" for a density histogram.}

\item{caption}{the figure caption.}

\item{margin}{set up the plot area margins.}

\item{\dots}{additional arguments for other methods.}
}
\value{
Information about the graph.
}
\description{
Create either a frequency or density histogram.
}
\details{
To set the x-axis range, you must specify numeric breaks that span the
complete range of \code{x}.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{histGram}.
}
\examples{
\dontrun{
# See for examples of histGram:
vignette(topic="ProbabilityPlots", package="smwrGraphs")
}
}
\references{
#'Helsel, D.R., and Hirsch, R.M., 2002, Statistical methods in water resources:
U.S. Geological Survey Techniques of Water-Resources Investigations, book 4,
chap. A3, 522 p.
}
\seealso{
\code{\link{ecdfPlot}}
}
\keyword{hplot}

