% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/histGram.R
\name{histGram}
\alias{histGram}
\alias{histGram.default}
\title{Histogram}
\usage{
histGram(x, breaks = "Sturges", Hist = list(), yaxis.range = c(NA, NA),
  ylabels = 7, xlabels = "Auto", xtitle = "", ytitle = "Auto",
  caption = "", margin = c(NA, NA, NA, NA), ...)

\method{histGram}{default}(x, breaks = "Sturges", Hist = list(type =
  "frequency", fill = FALSE, boundary = "lower", line.color = "black",
  fill.color = "gray80"), yaxis.range = c(NA, NA), ylabels = 7,
  xlabels = 7, xtitle = deparse(substitute(x)), ytitle = "Auto",
  caption = "", margin = c(NA, NA, NA, NA), ...)
}
\arguments{
\item{x}{a numeric vector to create the histogram}

\item{breaks}{any valid value for \code{\link{hist}}. See \bold{Details}.}

\item{Hist}{Controls for the histogram.}

\item{yaxis.range}{set the range for the y axis, the first value must be 0.}

\item{ylabels}{the approximate number of labels for the y axis.}

\item{xlabels}{the approximate number of labels for the x axis. The default value,
"Auto" sets labels that are aligned with the breaks.}

\item{xtitle}{x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption),
"Frequency" for a frequency histogram,
"Density" for a density histogram.}

\item{caption}{the figure caption.}

\item{margin}{set up the plot area margins.}

\item{\dots}{additional arguments for other methods.}
}
\value{
Information about the graph.
}
\description{
Create either a frequency or density histogram.
}
\details{
To set the x-axis range, you must specify numeric breaks that span the
complete range of \code{x}.

The components of \code{Hist}:
\describe{
\item{"type"}{The type of the histogram. Must be one of "frequency" for actual counts
in the bin, "density" for density in each bin, or "relative frequency" for percent in each bin.}
\item{"fill"}{Logical value, \code{TRUE} means each bin will be shaded with \code{fill.color}.}
\item{"boundary"}{Defines how values tied to bin limit boundaires are handled. If "upper,"
then the bin limit boundary is the upper limit of the range and values tied to that value
are placed in the bin corresponding to the upper limit of the boundary. If "lower," then
the bijn limit is the lower limit of the bin.}
\item{line.color}{The color of the lines around the bins.}
\item{fill.color}{The color the bins.}
}
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{histGram}.
}
\examples{
\dontrun{
set.seed(1)
Xbig <- rnorm(100)
histGram(Xbig, breaks=seq(-3, 3, by=.5), Hist=list(type="density"))
# For more details of histGram see
vignette(topic="ProbabilityPlots", package="smwrGraphs")
}
}
\references{
#'Helsel, D.R., and Hirsch, R.M., 2002, Statistical methods in water resources:
U.S. Geological Survey Techniques of Water-Resources Investigations, book 4,
chap. A3, 522 p.
}
\seealso{
\code{\link{ecdfPlot}}
}
\keyword{hplot}

