% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hull.R
\name{hull}
\alias{hull}
\title{Construct a Hull}
\usage{
hull(x, y, percent = 100, smooth = FALSE)
}
\arguments{
\item{x}{the x-coordinate data. Missing values are permitted, but ignored.}

\item{y}{the y-coordinate data. Missing values are permitted, but ignored.}

\item{percent}{the minimum percent to enclose.}

\item{smooth}{logical, if \code{TRUE}, then smooth the bounding hull.}
}
\value{
A list containing the x- and y-coordinates of the hull.
}
\description{
Construct an enclosing hull from x- and y-coordinate data.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
hull(X, Y)
# For examples of hull in graphs see
vignette(topic="GraphAdditions", package="smwrGraphs")
}
}
\seealso{
\code{\link{dataEllipse}}
}
\keyword{dplot}

