% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labelPoints.R
\name{labelPoints}
\alias{labelPoints}
\title{Label Points}
\usage{
labelPoints(x, y, labels, dir = "E", offset = 0.75, size = 8,
  color = "black", current = list(yaxis.log = FALSE, yaxis.rev = FALSE,
  xaxis.log = FALSE))
}
\arguments{
\item{x}{the x-axis data. Missing values are permitted, but ignored.}

\item{y}{the y-axis data. Missing values are permitted, but ignored.}

\item{labels}{the text labels, must be the same length as x and y. Missing
values are permitted, but ignored.}

\item{dir}{the direction relative to the point to place the label. See \bold{Details}.}

\item{offset}{the relative offset from the point. See \bold{Details}.}

\item{size}{character size in points.}

\item{color}{the color of the labels.}

\item{current}{the current plot controls. Typically, this would be the
output from one of the graph creation functions like \code{xyPlot}.}
}
\value{
A list containing \code{x}, \code{y}, and \code{labels}.
}
\description{
Labels points on a graph.
}
\details{
The value for \code{dir} can be of length one or the length of \code{x},
in which case, \code{dir} applies to the corresponding point. The value must be
"N," "NE," "E," "SE," "S," "SW," "W," or "NW" cooresponding to the compass
direction or "C" to center the label.

The value for \code{offset} can be of length one or the length of \code{x}, in
in which case, \code{offset} applies to the corresponding point. The value is relative
to the size of the test. The default value of \code{offset} is correct for creating a
text plot, where the text is centered on the value.
}
\note{
The current version does not have any method to automatically eliminate
overlapping labels. Several iterations may be required trying various values for
\code{dir} and possibly \code{offset} to produce non overlapping labels.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
setGD()
xyPlot(X, Y)
# Label the first point
labelPoints(X[1], Y[1], "First")
# For more details of labelPoints see
vignette(topic="GraphAdditions", package="smwrGraphs")
}
}
\seealso{
\code{\link{addAnnotation}},, \code{\link{xyPlot}}
}
\keyword{aplot}

