% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineWt.R
\name{lineWt}
\alias{frameWt}
\alias{lineWt}
\alias{stdWt}
\title{Line Weights}
\usage{
lineWt(x)

frameWt()

stdWt(x = 1)
}
\arguments{
\item{x}{for \code{lineWt}, one of "standard" (0.7 pt), "color" (0.8 pt),
"bold" (1.0 pt), or "hairline" (0.5 pt). For \code{stdWt}, a multiplier 
to set the line weight.}
}
\value{
The width of the line to set as the \code{lwd} graphics parameter.
}
\description{
Computes the weight, or width, of a line. Used primarily 
as a support function.
}
\seealso{
\code{\link{par}}
}
\keyword{dplot}

