% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logPretty.R
\name{logPretty}
\alias{logPretty}
\title{Pretty Axis}
\usage{
logPretty(x, hard = FALSE, labels = "Auto", style = "Auto",
  extend.pct = 0, extend.range = NA)
}
\arguments{
\item{x}{data defining the range to be plotted on the axis. Missing values
are permitted, but ignored.}

\item{hard}{logical, if \code{TRUE}, then use the minimum and maximum of
\code{x} as the fixed range of the axis, otherwise find "nice" limits.}

\item{labels}{either "Auto," which lets the function decide how many labels,
the approximate number of labels, or the actual numeric values of the labels.}

\item{style}{a character string indicating the style of the axis labels.
Valid values are "numeric,"
which forces the labels to be displayed as numbers; "scientific," which forces
the labels displayed using scientific notation; or "Auto" (the default), which
disaplays labels as numbers but switches to scientific notation for large ranges.
Only the first letter is required. Any invalid value will produce simply formatted
labels.}

\item{extend.pct}{extend the axis range by \code{extend.pct}. Only valid
when \code{hard} is FALSE.}

\item{extend.range}{required for naming consistency with other functions, not used.}
}
\value{
Information about the axis lables.
}
\description{
Constructs information for making a nicely formatted log-scale numeric axis.
A support function for creating logarithmic axes.
}
\seealso{
\code{\link{areaPlot}}, \code{\link{boxPlot}},
\code{\link{colorPlot}}, \code{\link{areaPlot}}, \code{\link{dotPlot}},
\code{\link{ecdfPlot}}, \code{\link{probPlot}}, \code{\link{qqPlot}},
\code{\link{scalePlot}}, \code{\link{splomPlot}}, \code{\link{timePlot}},
\code{\link{xyPlot}}
}
\keyword{dplot}

