% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paraSpline.R
\name{paraSpline}
\alias{paraSpline}
\title{Parametric Spline}
\usage{
paraSpline(x, y, n)
}
\arguments{
\item{x}{the x-coordinate data. Missing values are not permitted.}

\item{y}{the y-coordinate data. Missing values are not permitted.}

\item{n}{The number of points in the output parametric spline fit.}
}
\value{
A list containing the components \code{x} and \code{y}, which are
the coordinates of the parametric spline.
}
\description{
Constructs a parametric interpolating spline for x and y data. The x data
are not required to be strictly increasing. Used as a support function.
}
\examples{
paraSpline(c(1,2,3), c(0,1,0), n=5)
}
\keyword{dplot}

