% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderBXP.R
\name{renderBoxPlot}
\alias{renderBXP}
\alias{renderBoxPlot}
\title{Box Plot}
\usage{
renderBoxPlot(xtoplot, stats, Box, explan, expz, yaxis.log = FALSE,
  yrange = c(NA, NA), xrange = range(xtoplot) + c(-1, 1),
  ylabels = "Auto", xlabels = "Auto", xlabels.rotate = FALSE,
  xtitle = "", ytitle = "", caption = "", margin = c(NA, NA, NA, NA))

renderBXP(x, width, z, draw.RL = TRUE, fill = "none")
}
\arguments{
\item{xtoplot}{the x-axis locations for each boxplot.}

\item{stats}{a list containing the statistics for the boxplots.}

\item{Box}{a list containing the control info for the boxplots.}

\item{explan}{a list containing the information for an explanation.}

\item{expz}{a list containing the information for an explanation of the
boxplot.}

\item{yaxis.log}{logical, if \code{TRUE}, then use a log transform for 
the data and the y-axis.}

\item{yrange}{set the y-axis range.}

\item{xrange}{set the x-axis range.}

\item{ylabels}{either "Auto," the approximate number of labels, or the
actual labels to use for the y-axis.}

\item{xlabels}{either "Auto" or the x-axis labels for each boxplot.}

\item{xlabels.rotate}{logical, if \code{TRUE}, then rotate the x-axis 
labels by 90 degrees.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{the parameters of the margin of the plot area.}

\item{x}{the x-coordinate for the box.}

\item{width}{the width of the box in x-axis units.}

\item{z}{a list containing the statistics for the individual boxplot.}

\item{draw.RL}{logical, if \code{TRUE}, then draw the reporting level
for the individual boxplot.}

\item{fill}{a character string describing the fill color for the box or
"none" for no fill.}
}
\value{
Information about the graph.
}
\description{
Draws a box plot (support functions).
}
\note{
The function \code{renderBXP} draws a single boxplot in a graph. The
function \code{renderBoxPlot} is called from each method function for
\code{boxPlot} to produce the boxplot.
}
\seealso{
\code{\link{boxPlot}}
}
\keyword{aplot}

