% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalePlot.R
\name{scalePlot}
\alias{scalePlot}
\title{Scale Plot}
\usage{
scalePlot(x, y, scale = 1, Plot = list(name = "", what = "lines", type
  = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA,
  NA), ylabels = 7, xlabels = 7, xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA,
  NA, NA))
}
\arguments{
\item{x}{the x-axis data.}

\item{y}{the y-axis data.}

\item{scale}{the y/x ratio. See \bold{Details}.}

\item{Plot}{parameters defining the characteristics of the plot. See
\code{\link{setPlot}} for a description of the parameters.}

\item{yaxis.log}{logical, if \code{TRUE}, then log-transform the y axis.}

\item{yaxis.rev}{logical, if \code{TRUE}, then reverse the y axis.}

\item{yaxis.range}{set the range of the y axis. See \bold{Details}.}

\item{xaxis.log}{logical, if \code{TRUE}, then log-transform the x axis.}

\item{xaxis.range}{set the range of the x axis. See \bold{Details}.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{set the plot area margins, in units of lines of text. Generally
all NA or the output from \code{setGraph} if appropriate.}
}
\value{
Information about the graph.
}
\description{
Produces a graph with a fixed aspect ratio for the x- and y-axes.
}
\details{
The \code{scale} argument sets the scaling ratio of the y-axis to the
x-axis. For latitude and longitude data, set the scale to
1/cos(midlat/180*pi), where midlat is the midrange of the latitude.

For linear axes, the range can be set to virtually any pair of values. For 
log axes, the choice of range is more resticted---for less than one log-cycle,
powers of whole numbers can be used; from 1 to about 3 log cycles, the choces 
should be powers of 3 or 10; and for more than 3 log cycles, the range sould be 
expressed only in powers of 10.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{scalePlot}.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(32)
Y <- X + rnorm(32)
setGD()
scalePlot(X, Y, Plot=list(what="points", size=0.05))
# For more details of scalePlot see
vignette(topic="LineScatter", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{xyPlot}}
}
\keyword{hplot}
