% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/seriesPlot.R
\name{seriesPlot}
\alias{seriesPlot}
\alias{seriesPlot.default}
\title{Series Plot}
\usage{
seriesPlot(x, SeasonLine = list(name = "", what = "vertical", color =
  "black"), SeasonPoint = list(name = "", what = "points", symbol = "circle",
  filled = TRUE, size = 0.09, color = "black"), yaxis.log = FALSE,
  yaxis.range = c(NA, NA), ylabels = 7, xlabels, xtitle = "",
  ytitle = "", caption = "", margin = c(NA, NA, NA, NA), ...)

\method{seriesPlot}{default}(x, SeasonLine = list(name = "", what =
  "vertical", color = "black"), SeasonPoint = list(name = "", what = "points",
  symbol = "circle", filled = TRUE, size = 0.09, color = "black"),
  yaxis.log = FALSE, yaxis.range = c(NA, NA), ylabels = 7,
  xlabels = frequency(x), xtitle = "", ytitle = deparse(substitute(x)),
  caption = "", margin = c(NA, NA, NA, NA), ...)
}
\arguments{
\item{x}{data that can be treated as a regularly-spaced time series. Missing
values are permitted, but result in missing seasons.}

\item{SeasonLine}{control parameters of the lines in the plot. See
\bold{Details}.}

\item{SeasonPoint}{control parameters of the points in the plot. See
\bold{Details}.}

\item{yaxis.log}{logical, if \code{TRUE}, then log-transform the y axis.}

\item{yaxis.range}{set the range of the y axis. See \bold{Details}.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set the x-axis labels and number of seasons when \code{x} is
a simple numeric vector, may be a single numeric value indicating the number
of seasons in \code{x} or a vector of the names of the seasons. See
\code{\link{namePretty}} for details. If \code{X} is a time-series object,
then the labels are set to the frequency characteristic of \code{x}.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{caption}{the figure caption.}

\item{margin}{the parameters of the margin of the plot area.}

\item{\dots}{any additional arguments required for specific methods.}
}
\value{
Information about the graph.
}
\description{
Produce a plot.
}
\details{
The argument \code{what} for \code{SeasonLine} must be either "lines" or
"vertical."  See \code{\link{monthplot}} for more information.

The argument \code{what} for \code{SeasonPoint} can be set to "none" to suppress
drawing of that feature.

For linear axes, the range can be set to virtually any pair of values. For
log axes, the choice of range is more resticted---for less than one log-cycle,
powers of whole numbers can be used; from 1 to about 3 log cycles, the choces
should be powers of 3 or 10; and for more than 3 log cycles, the range sould be
expressed only in powers of 10.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{seriesPlot}.
}
\examples{
\dontrun{
# See for examples of seriesPlot:
vignette(topic="LineScatter", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{monthplot}},
\code{\link{seasonPlot}}
}
\keyword{hplot}

