% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setColor}
\alias{setColor}
\title{Colors}
\usage{
setColor(Color)
}
\arguments{
\item{Color}{any kind of data that might be interpreted as a color.}
}
\value{
The values in \code{Color} converted to a value that could be
interpreted as a color.
}
\description{
Checks or converts anything to valid colors (support function).
}
\seealso{
\code{link{colors}}, \code{link{rainbow}}
}
\keyword{dplot}

