% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDefaults.R
\name{setDefaults}
\alias{setDefaults}
\title{Default Values}
\usage{
setDefaults(current = list(), ...)
}
\arguments{
\item{current}{the control parameters specified in the call to the
high-level graphing function.}

\item{\dots}{the default values for each name required for the control
parameters.}
}
\value{
The control parameters with defaults substituted for missing names.
}
\description{
Sets the default values for plot control lists (support function).
}
\keyword{dplot}
