% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setExplan.R
\name{setExplan}
\alias{setExplan}
\title{Explanation}
\usage{
setExplan(current, old = NULL)
}
\arguments{
\item{current}{the current plot information.}

\item{old}{existing explanation information.}
}
\value{
A list having four components:\cr
\item{text}{a list having two components:}
\tabular{ll}{
 text \tab the description in the explanation, derived from
the \code{name} component in \code{Plot}\cr
cex \tab the size of the \code{text} to write in the explanation\cr}
\item{lines}{a list having seven components and controlling both lines and
points:}
\tabular{ll}{
type \tab the type of plot (specifying points, lines, and so forth)\cr
lwd \tab the line weight\cr
lty \tab the line type\cr
pch \tab the symbol\cr
csi \tab the size of the symbol in inches\cr
cex \tab the the size of the symbol relative to character size\cr
col \tab the color of the plot\cr}
\item{areas}{a list having two components:}
\tabular{ll}{
fill \tab the color of the fill area\cr
borders \tab the color of the border\cr}
\item{current}{a list like \code{current} with the defaults set}
Each entry in \code{text} must have a corresponding entry in \code{lines} and
\code{areas}.
}
\description{
Adds the current plot information to explanation information (support
function).
}
\seealso{
\code{\link{setPlot}}
}
\keyword{dplot}

