% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setGraph}
\alias{setGraph}
\title{Set Graph}
\usage{
setGraph(graphNum, layout, noTicks = NULL)
}
\arguments{
\item{graphNum}{the number of the graph.}

\item{layout}{the layout of the graph.}

\item{noTicks}{suppress ticks on a specific axis.}
}
\value{
The parameters of the margin of the plot area.
}
\description{
Sets ups page for a specific grtaph
}
\note{
This function is called using the information generated by
\code{setLayout}. It invisibly sets up the graphics device for the next
graph.
}
\seealso{
\code{\link{setLayout}}
}
\keyword{dplot}

