% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setRtMargin}
\alias{setRtMargin}
\alias{setTopMargin}
\title{Set Secondary Margin}
\usage{
setRtMargin(y, margin = c(NA, NA, NA, NA), right.labels = 7,
  right.log = FALSE, right.range = c(NA, NA))

setTopMargin(margin = c(NA, NA, NA, NA))
}
\arguments{
\item{y}{the secondary y-axis data to be plotted, missing values are
permitted and are ignored.}

\item{margin}{incomplete plot margin specification, generally computed by
\code{setGraph}.}

\item{right.labels}{set up right-axis labels; the approximate number of
labels.}

\item{right.log}{logical: log transform right axis?}

\item{right.range}{set right-axis range.}
}
\value{
The updated margin; only the right margin value is changed.
}
\description{
Set the right or top margin for graphs with secondary axes. Used After
setting up the graphics environment, but before the call the the high-level
graphics function to allocate space for an additional label and title.
}
\details{
The values for \code{right.labels}, \code{right.log}, and \code{right.range}
should be set exactly as in the call to \code{addXY}.
}
\seealso{
\code{\link{setLayout}}, \code{\link{setGraph}},
\code{\link{addXY}}
}
\keyword{dplot}

