% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setRtMargin.R
\name{setRtMargin}
\alias{setRtMargin}
\alias{setTopMargin}
\title{Set Secondary Margin}
\usage{
setRtMargin(y, margin = c(NA, NA, NA, NA), right.labels = 7,
  right.log = FALSE, right.range = c(NA, NA))

setTopMargin(margin = c(NA, NA, NA, NA))
}
\arguments{
\item{y}{the secondary y-axis data to be plotted, missing values are
permitted and are ignored.}

\item{margin}{incomplete plot margin specification, generally computed by
\code{setGraph}.}

\item{right.labels}{set up right-axis labels; the approximate number of
labels.}

\item{right.log}{logical, if \code{TRUE}, then log transform right axis.}

\item{right.range}{set right-axis range.}
}
\value{
The updated margin; only the right margin value is changed.
}
\description{
Sets the right or top margin for graphs with secondary axes. Used After
setting up the graphics environment, but before the call to the high-level
graphics function to allocate space for an additional label and title.
}
\details{
The values for \code{right.labels}, \code{right.log}, and \code{right.range}
should be set exactly as in the call to \code{addXY}.

The \code{margin} is a numeric vector of length 4 specifying the bottom, left, top,
and right margins around the plot, as decriibed in by the \code{mar} option in
\code{\link{par}}. The function \code{setTopMargin} only changes the third value and
the function \code{setRtMargin} only changes the fourth value. The value for
\code{margin} is typically the output from \code{setGraph} or the defaults for
these functions.
}
\examples{
\dontrun{
# See for examples of setRtMargin:
demo(topic="RightAxisExample", package="smwrGraphs")
# See for examples of setTopMargin:
demo(topic="TopAxisExample", package="smwrGraphs")
}
}
\seealso{
\code{\link{setLayout}}, \code{\link{setGraph}},
\code{\link{addXY}}
}
\keyword{dplot}

