% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strip.blanks}
\alias{strip.blanks}
\title{Remove Spaces}
\usage{
strip.blanks(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
A vector like \code{x}, but with leading and trailing spaces removed
from each element.
}
\description{
Remove surrounding spaces from text.
}
\examples{
strip.blanks("   keep me   ")
}
\seealso{
\code{link{sub}}
}
\keyword{manip}

