% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surfacePlot.R
\name{surfacePlot}
\alias{surfacePlot}
\title{Surface plot}
\usage{
surfacePlot(pre, z.color = "lightblue", Surface = list(name = "", lineColor
  = "black", levels = 20, ramp = "coolWarm"), xtitle = "", ytitle = "",
  ztitle = "", margin = c(NA, NA, NA, NA), caption = "")
}
\arguments{
\item{pre}{the output from \code{preSurface}.}

\item{z.color}{the surface color. Can be either a character string indicating the
color of the surface, or a numeric matrix. If a matrix, then must have
\code{length(x)} rows and \code{length(y)} columns or \code{length(x) - 1} rows
and \code{length(y)-1} columns. Missing values are permitted, but result
in blank areas on the surface. See \bold{Details}.}

\item{Surface}{control parameters for the surface. See \bold{Details}.}

\item{xtitle}{the x-axis title (also called x-axis caption).}

\item{ytitle}{the y-axis title (also called y-axis caption).}

\item{ztitle}{the z-axis title (also called z-axis caption).}

\item{margin}{set up the plot area margins. To allocate space for a graph title,
set the third value to 1.5, otherwise all values should be \code{NA} or 0.}

\item{caption}{the figure caption.}
}
\value{
Information about the graph.
}
\description{
Creates a surface plot to show three-dimensional data or a colored surface plot
to show four-dimensional data.
}
\details{
If \code{z.color} is a numeric matrix, then the values represent the average
surface of the grid defined by \code{x} and \code{y}, that is it represents the
value one-half way between each value of \code{x} and one-half way between each
value of \code{y}, the mid point. If it has the same dimensions as \code{z.surf},
then the data are resampled by averaging each of the four corners to compute the
mid point value. The values of \code{z.color} are assigned using the controls
in \code{Surface}.\cr
The \code{Surface} argument must be a tagged list with these components:
\describe{
 \item{name}{the name of \code{z.color} used in the explanation.}
 \item{lineColor}{the color for each line on the surface. If "none," then
 lines are not drawn.}
 \item{levels}{the levels of the surface colors. May be either a single numeric
 value that indicates the approximate number of levels, or a vector that
 indicates the exact breaks in the levels.}
 \item{ramp}{the name of the color ramp. May be "gray" or "grey" for a gray
 scale or the prefix name of a function that creates a range of colors, see
 \code{\link{coolWarm.colors}} for examples.}
}
}
\note{
A call must be made to \code{setPage} or \code{setPDF} to set up the
graphics environment before calling \code{surfacePlot}.
}
\examples{
\dontrun{
# See for examples of surfacePlot:
vignette(topic="GraphGallery", package="smwrGraphs")
}
}
\seealso{
\code{\link{setPage}}, \code{\link{preSurface}},
\code{\link[graphics]{persp}}
}
\keyword{hplot}

