% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePretty.R
\name{timePretty}
\alias{timePretty}
\title{Pretty Axis}
\usage{
timePretty(x, labels = "Auto")
}
\arguments{
\item{x}{time difference data}

\item{labels}{either "Auto," which lets the function decide how many labels,
the approximate number of labels, or the actual labels to use.}
}
\value{
Information about the axis labels.
}
\description{
Constructs information for making a nicely formatted date/time axis.
}
\seealso{
\code{\link{timePlot}}
}
\keyword{dplot}

