% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timePretty}
\alias{timePretty}
\title{Pretty Axis}
\usage{
timePretty(x, labels = "Auto")
}
\arguments{
\item{x}{time difference data}

\item{labels}{either "Auto," which lets the function decide how many labels,
the approximate number of labels, or the actual labels to use.}
}
\value{
information about the axis labels.
}
\description{
Construct information for making a a nicely formatted date axis.
}
\seealso{
\code{\link{timePlot}}
}
\keyword{dplot}

