% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transData.R
\name{transData}
\alias{transData}
\title{Transform Data}
\usage{
transData(data, logT = FALSE, revT = FALSE, trans = as.vector,
  transarg = NULL)
}
\arguments{
\item{data}{data for axis}

\item{logT}{logical, if \code{TRUE}, then log transform \code{data}}

\item{revT}{logical, if \code{TRUE}, then reverse \code{data} to match the axis.}

\item{trans}{arbitrary transform function to apply to \code{data}.}

\item{transarg}{list of arguments to \code{trans}.}
}
\value{
A vector like \code{data} transformed to plot correctly on an axis.
}
\description{
Transforms numeric data to match any axis scaling (support function).
}
\seealso{
\code{\link{transPlot}}, \code{\link{probPlot}}
}
\keyword{dplot}

